feat(deps): pin xet-core to commit SHA and bump all dependencies
Pin xet-core dependencies to commit 40f9530 for reproducibility. Bump all external dependencies to latest compatible versions: tokio 1.52.3, tonic/prost 0.14.6, reqwest 0.13.4, clap 4.6.1, and others. Add security audit step to CI workflow. Add CSI integration tests and Windows cross-compile check to CI.
