+++ src/virtual_fs/mod.rs
@@ -84,11 +84,6 @@ pub struct VfsConfig {
     pub direct_io: bool,
     pub flush_debounce: Duration,
     pub flush_max_batch_window: Duration,
-    /// Max time the SIGTERM shutdown drain may spend flushing dirty data before
-    /// abandoning it to guarantee the process exits within the pod's termination
-    /// grace period. Must be < terminationGracePeriodSeconds, otherwise a slow
-    /// Hub/CAS backend wedges the FUSE connection and strands the pod.
-    pub flush_shutdown_timeout: Duration,
     /// 0 disables the LRU evictor.
     pub inode_soft_limit: usize,
     pub lru_sweep_interval: Duration,
@@ -116,8 +111,6 @@ pub struct VfsConfig {
 /// inode.size between the file truncation and the metadata update.
 pub struct VirtualFs {
     runtime: tokio::runtime::Handle,
-    /// Upper bound on the SIGTERM flush drain (see `VfsConfig`).
-    flush_shutdown_timeout: Duration,
     hub_client: Arc<dyn HubOps>,
     xet_sessions: Arc<dyn XetOps>,
     /// Staging area + per-inode locks. The per-inode locks are always
@@ -179,3474 +172,15 @@ pub struct VirtualFs {
     file_cache: Option<Arc<FileCache>>,
 }
 
-impl VirtualFs {
-    pub fn new(
-        runtime: tokio::runtime::Handle,
-        hub_client: Arc<dyn HubOps>,
-        xet_sessions: Arc<dyn XetOps>,
-        staging_dir: Option<StagingDir>,
-        file_cache: Option<Arc<FileCache>>,
-        overlay_backing: Option<OverlayBacking>,
-        config: VfsConfig,
-    ) -> Arc<Self> {
-        let inodes = Arc::new(RwLock::new(InodeTable::new(config.inode_soft_limit > 0)));
-        let negative_cache = Arc::new(RwLock::new(HashMap::new()));
+mod core;
+mod handle;
+mod lookup;
+mod attr;
+mod open;
+mod io;
+mod dir;
+mod rename;
 
-        let staging = Arc::new(StagingCoordinator::new(staging_dir));
-        let overlay_backing = overlay_backing.map(Arc::new);
-        let overlay = overlay_backing.is_some();
-
-        // Staging GC orders eviction by `last_touched`. When the inode evictor
-        // is off (`soft_limit==0`) the touch hook is a no-op, so arm it
-        // explicitly when a staging budget is configured — otherwise the GC
-        // would fall back to arbitrary HashMap order.
-        if staging.dir().is_some_and(|sd| sd.has_budget()) {
-            inodes.read().expect("inodes poisoned").enable_touch_tracking();
-        }
-
-        // Overlay mode keeps writes local: skip the flush pipeline entirely.
-        let flush_manager = if !config.read_only && config.advanced_writes && !overlay {
-            staging.dir().expect("--advanced-writes requires a staging directory");
-            Some(flush::FlushManager::new(
-                xet_sessions.clone(),
-                staging.clone(),
-                hub_client.clone(),
-                inodes.clone(),
-                &runtime,
-                config.flush_debounce,
-                config.flush_max_batch_window,
-            ))
-        } else {
-            None
-        };
-
-        // Create open_files before poll task so we can share with it
-        let open_files: Arc<RwLock<HashMap<u64, OpenFile>>> = Arc::new(RwLock::new(HashMap::new()));
-
-        // Spawn remote change polling task (if interval > 0)
-        let invalidator: Invalidator = Arc::new(OnceLock::new());
-        let poll_handle = if config.poll_interval_secs > 0 {
-            let bg_hub = hub_client.clone();
-            let bg_inodes = inodes.clone();
-            let bg_neg_cache = negative_cache.clone();
-            let bg_invalidator = invalidator.clone();
-            let interval = Duration::from_secs(config.poll_interval_secs);
-            // Clamp to >= 1 in case a caller (library use) constructs VfsConfig directly.
-            let listing_concurrency = config.poll_listing_concurrency.max(1);
-
-            Some(runtime.spawn(Self::poll_remote_changes(
-                bg_hub,
-                bg_inodes,
-                bg_neg_cache,
-                bg_invalidator,
-                interval,
-                listing_concurrency,
-            )))
-        } else {
-            None
-        };
-
-        let entry_invalidator: EntryInvalidator = Arc::new(OnceLock::new());
-        let vfs = Arc::new(Self {
-            runtime,
-            flush_shutdown_timeout: config.flush_shutdown_timeout,
-            hub_client,
-            xet_sessions,
-            staging,
-            overlay_backing,
-            read_only: config.read_only,
-            // Overlay implies advanced_writes (random writes via local backing file).
-            advanced_writes: config.advanced_writes || overlay,
-            inode_table: inodes,
-            open_files,
-            next_file_handle: AtomicU64::new(1),
-            uid: config.uid,
-            gid: config.gid,
-            negative_cache,
-            dir_loading_locks: Mutex::new(HashMap::new()),
-            pending_commits: Mutex::new(HashMap::new()),
-            flush_manager,
-            poll_handle: Mutex::new(poll_handle),
-            invalidator,
-            entry_invalidator,
-            lru_handle: Mutex::new(None),
-            metadata_ttl: config.metadata_ttl,
-            serve_lookup_from_cache: config.serve_lookup_from_cache,
-            filter_os_files: config.filter_os_files,
-            direct_io: config.direct_io,
-            file_cache,
-        });
-
-        // Set root inode mtime and ownership (repos use the last commit date).
-        {
-            let mut inodes = vfs.inode_table.write().expect("inodes poisoned");
-            if let Some(root) = inodes.get_mut(inode::ROOT_INODE) {
-                root.mtime = vfs.hub_client.default_mtime();
-                root.atime = root.mtime;
-                root.uid = vfs.uid;
-                root.gid = vfs.gid;
-            }
-        }
-
-        // Pre-load root directory so `ls /mount` is instant. The LRU cap is
-        // armed at `InodeTable::new(soft_limit)` above, so a flat root with
-        // thousands of direct children won't blow past the budget here.
-        // Subdirectories are lazy-loaded on first access.
-        if let Err(e) = vfs.runtime.block_on(vfs.ensure_children_loaded(inode::ROOT_INODE)) {
-            error!("Failed to pre-load root directory: errno={}", e);
-        }
-
-        // Spawn LRU evictor. `Weak` so the task exits when the outer Arc is
-        // dropped; still aborted in shutdown() for determinism.
-        if config.inode_soft_limit > 0 {
-            let weak = Arc::downgrade(&vfs);
-            let soft_limit = config.inode_soft_limit;
-            let sweep_interval = config.lru_sweep_interval;
-            let handle = vfs
-                .runtime
-                .spawn(Self::lru_sweep_loop(weak, soft_limit, sweep_interval));
-            *vfs.lru_handle.lock().expect("lru_handle poisoned") = Some(handle);
-            info!(
-                "LRU evictor enabled: soft_limit={} inodes, sweep every {:?}",
-                soft_limit, sweep_interval
-            );
-        }
-
-        vfs
-    }
-    /// True when the VFS is in overlay mode (writes stay local, no remote
-    /// mutations). Derived from the presence of an overlay backing.
-    fn overlay(&self) -> bool {
-        self.overlay_backing.is_some()
-    }
-
-    /// Splice local overlay entries into the inode table for `parent_ino`,
-    /// overriding remote entries on name conflict (kind mismatch removes the
-    /// remote inode first, then re-inserts as local). No-op when not in
-    /// overlay mode.
-    fn merge_overlay_entries(&self, inodes: &mut InodeTable, parent_ino: u64) {
-        let Some(overlay) = &self.overlay_backing else {
-            return;
-        };
-        let dir_path = inodes.get(parent_ino).map(|e| e.full_path.clone()).unwrap_or_default();
-        let Ok(entries) = overlay.read_dir(&dir_path) else {
-            return;
-        };
-        for entry in entries {
-            if entry.is_symlink || (self.filter_os_files && is_os_junk(&entry.name)) {
-                continue;
-            }
-            let kind = if entry.is_dir {
-                InodeKind::Directory
-            } else {
-                InodeKind::File
-            };
-            let full_path = inode::child_path(&dir_path, &entry.name);
-            // If existing entry has a different kind (e.g. remote file vs
-            // local dir), remove it first so local wins cleanly.
-            let conflict = inodes
-                .lookup_child(parent_ino, &entry.name)
-                .filter(|e| e.kind != kind)
-                .map(|e| e.inode);
-            if let Some(old_ino) = conflict {
-                inodes.remove(old_ino);
-            }
-            let ino = inodes.insert(
-                parent_ino,
-                entry.name,
-                full_path,
-                kind,
-                entry.size,
-                entry.mtime,
-                None,
-                entry.mode,
-                self.uid,
-                self.gid,
-            );
-            // Local overrides remote: clear remote identity, mark dirty.
-            if let Some(e) = inodes.get_mut(ino) {
-                e.size = entry.size;
-                e.mtime = entry.mtime;
-                e.mode = entry.mode;
-                e.xet_hash = None;
-                if !e.is_dirty() {
-                    e.set_dirty();
-                }
-                if kind == InodeKind::Directory {
-                    e.children_loaded_at = None;
-                }
-            }
-        }
-    }
-
-    /// In overlay mode, signal that `full_path` exists locally and a remote
-    /// HEAD/list probe must be skipped to preserve local-overrides-remote.
-    /// On hit, re-runs `merge_overlay_entries` on the parent so the entry
-    /// shows up in the inode table even when it was added to the backing
-    /// after the parent's listing was loaded.
-    fn overlay_lookup_takes_precedence(&self, parent: u64, full_path: &str) -> VirtualFsResult<bool> {
-        let Some(overlay) = &self.overlay_backing else {
-            return Ok(false);
-        };
-        let exists = overlay.exists(full_path).map_err(|e| {
-            error!("Failed to stat overlay path {}: {}", full_path, e);
-            libc::EIO
-        })?;
-        if exists {
-            let mut inodes = self.inode_table.write().expect("inodes poisoned");
-            self.merge_overlay_entries(&mut inodes, parent);
-        }
-        Ok(exists)
-    }
-
-    /// True if the entry is a clean remote entry that overlay mode treats as immutable.
-    fn is_overlay_immutable(&self, entry: &inode::InodeEntry) -> bool {
-        self.overlay() && !entry.is_dirty()
-    }
-
-    /// Set the kernel cache invalidation callback. Called after mount setup
-    /// so the poll loop can actively invalidate stale inodes on remote changes.
-    /// First call wins; later calls are no-ops.
-    pub fn set_invalidator(&self, f: InvalidatorFn) {
-        let _ = self.invalidator.set(f);
-    }
-
-    /// First call wins; later calls are no-ops.
-    pub fn set_entry_invalidator(&self, f: EntryInvalidatorFn) {
-        let _ = self.entry_invalidator.set(f);
-    }
-
-    async fn lru_sweep_loop(weak: std::sync::Weak<Self>, soft_limit: usize, interval: Duration) {
-        loop {
-            tokio::time::sleep(interval).await;
-            let Some(vfs) = weak.upgrade() else { return };
-            let table_len = vfs.inode_table.read().expect("inodes poisoned").len();
-            let evicted = vfs.lru_evict_sweep(soft_limit).await;
-            if evicted > 0 || table_len > soft_limit {
-                info!(
-                    "lru_sweep: table={} soft_limit={} evicted={}",
-                    table_len, soft_limit, evicted
-                );
-            }
-        }
-    }
-
-    /// Candidates are collected under a read lock, then released before
-    /// invoking `cb` — `cb` writes to the FUSE notify channel and returns
-    /// `false` on EAGAIN/ENOMEM, which acts as the natural batch throttle.
-    ///
-    /// After `inval_entry`, we also evict our own table entry if `evict_if_safe`
-    /// lets us: when the inode was materialized from a readdir the kernel
-    /// never cached a dentry for it (nlookup stays at 0), so no `forget()`
-    /// will ever come back. Waiting on one leaks the entry forever.
-    ///
-    /// The invalidation loop runs on `spawn_blocking` in chunks of
-    /// `INVAL_BATCH_SIZE` with a `INVAL_BATCH_PAUSE` between chunks so the
-    /// kernel's reverse-notify path doesn't starve concurrent FUSE ops.
-    async fn lru_evict_sweep(&self, soft_limit: usize) -> usize {
-        let candidates = {
-            let inodes = self.inode_table.read().expect("inodes poisoned");
-            let len = inodes.len();
-            if len <= soft_limit {
-                return 0;
-            }
-            inodes.lru_candidates(len - soft_limit)
-        };
-        if candidates.is_empty() {
-            return 0;
-        }
-        if self.entry_invalidator.get().is_none() {
-            return 0;
-        }
-
-        let mut to_evict = Vec::with_capacity(candidates.len());
-        let mut iter = candidates.into_iter();
-        loop {
-            let chunk: Vec<(u64, u64, Arc<str>)> = iter.by_ref().take(INVAL_BATCH_SIZE).collect();
-            if chunk.is_empty() {
-                break;
-            }
-            // A short chunk means the iterator was exhausted while filling
-            // it — no need to pause after, there is nothing more to process.
-            let is_last = chunk.len() < INVAL_BATCH_SIZE;
-            let invalidator = self.entry_invalidator.clone();
-            let result = tokio::task::spawn_blocking(move || {
-                let Some(invalidate_entry) = invalidator.get() else {
-                    return (Vec::new(), true);
-                };
-                let mut evicted = Vec::with_capacity(chunk.len());
-                for (ino, parent, name) in chunk {
-                    // `invalidate_entry` returns false when the FUSE notify
-                    // channel is saturated (EAGAIN/ENOMEM) — backpressure,
-                    // stop the sweep.
-                    if !invalidate_entry(parent, &name) {
-                        return (evicted, true);
-                    }
-                    evicted.push(ino);
-                }
-                (evicted, false)
-            })
-            .await;
-            let stop = match result {
-                Ok((mut chunk_inos, stop)) => {
-                    to_evict.append(&mut chunk_inos);
-                    stop
-                }
-                Err(e) => {
-                    warn!("lru_sweep: invalidation batch joined with error: {e}");
-                    true
-                }
-            };
-            if stop || is_last {
-                break;
-            }
-            // Pause so the kernel can drain reverse-notify and concurrent
-            // FUSE ops can make progress.
-            tokio::time::sleep(INVAL_BATCH_PAUSE).await;
-        }
-
-        if to_evict.is_empty() {
-            return 0;
-        }
-        // Evict in chunks so the write lock is released between batches —
-        // a single huge batch could hold it for hundreds of ms, stalling
-        // every concurrent lookup / readdir / forget. `evict_batch_if_safe`
-        // also collapses the per-eviction `parent.children.retain()` into
-        // one pass per touched parent, which matters when many evicted
-        // siblings share the same dir.
-        const EVICT_CHUNK: usize = 4096;
-        let mut total_evicted = 0;
-        for chunk in to_evict.chunks(EVICT_CHUNK) {
-            let evicted_inos = {
-                let mut inodes = self.inode_table.write().expect("inodes poisoned");
-                inodes.evict_batch_if_safe(chunk)
-            };
-            // Reclaim per-inode staging files, mirroring forget()/release().
-            for ino in &evicted_inos {
-                self.drop_staging(*ino);
-            }
-            total_evicted += evicted_inos.len();
-        }
-        total_evicted
-    }
-
-    /// Graceful shutdown: abort polling, drain flush queue, wait for completion.
-    pub fn shutdown(&self) {
-        info!("Shutting down VFS, flushing pending writes...");
-        // Abort background tasks.
-        if let Some(handle) = self.poll_handle.lock().expect("poll_handle poisoned").take() {
-            handle.abort();
-        }
-        if let Some(handle) = self.lru_handle.lock().expect("lru_handle poisoned").take() {
-            handle.abort();
-        }
-        // Flush all dirty files + queued deletes.
-        if let Some(fm) = &self.flush_manager {
-            let dirty = self.inode_table.read().expect("inodes poisoned").dirty_inos();
-            fm.shutdown(dirty, &self.runtime, self.flush_shutdown_timeout);
-        }
-        info!("Flush loop finished, VFS shut down.");
-    }
-
-    // ── Helpers ─────────────────────────────────────────────────────────
-
-    fn make_vfs_attr(&self, entry: &InodeEntry) -> VirtualFsAttr {
-        let perm = if self.read_only {
-            match entry.kind {
-                InodeKind::File => 0o444,
-                InodeKind::Directory => 0o555,
-                InodeKind::Symlink => 0o777,
-            }
-        } else {
-            match entry.kind {
-                InodeKind::Symlink => 0o777,
-                _ => entry.mode,
-            }
-        };
-
-        VirtualFsAttr {
-            ino: entry.inode,
-            size: entry.size,
-            blocks: entry.size.div_ceil(BLOCK_SIZE as u64),
-            mtime: entry.mtime,
-            atime: entry.atime,
-            ctime: entry.ctime,
-            kind: entry.kind,
-            perm,
-            nlink: entry.nlink,
-            uid: entry.uid,
-            gid: entry.gid,
-        }
-    }
-
-    /// Revalidate a remote file by checking the Hub for metadata changes.
-    /// Skips HEAD if the inode was validated within `metadata_ttl`.
-    /// If the file's xet_hash changed, updates the inode and invalidates kernel cache.
-    /// If the file was deleted (404), removes the inode.
-    /// On network errors, silently returns (graceful degradation).
-    async fn revalidate_file(&self, ino: u64, full_path: &str, current_hash: Option<&str>, current_etag: Option<&str>) {
-        // When serve_lookup_from_cache is true, skip HEAD if recently validated.
-        // When false (minimal mode), always HEAD on every lookup.
-        if self.serve_lookup_from_cache {
-            let inodes = self.inode_table.read().expect("inodes poisoned");
-            if let Some(entry) = inodes.get(ino)
-                && let Some(last) = entry.last_revalidated
-                && last.elapsed() < self.metadata_ttl
-            {
-                return;
-            }
-        }
-
-        let remote = match self.hub_client.head_file(full_path).await {
-            Ok(r) => r,
-            Err(e) => {
-                debug!("head_file({}) failed, using cached: {}", full_path, e);
-                return;
-            }
-        };
-
-        // Collect kernel invalidations to issue after we drop the inode_table
-        // write lock. `inval_inode` issues a synchronous writev to /dev/fuse that
-        // ends up in `invalidate_inode_pages2_range` on the kernel side, which
-        // blocks on per-folio waits. Calling it under the global write lock can
-        // stall the whole VFS (and has been observed to deadlock production
-        // pods). Apply the mutation, drop the lock, then notify.
-        let mut to_invalidate: Vec<u64> = Vec::new();
-        match remote {
-            None => {
-                // File deleted remotely → remove from inode table
-                info!("Remote deletion detected via HEAD: {}", full_path);
-                let mut inodes = self.inode_table.write().expect("inodes poisoned");
-                if let Some(entry) = inodes.get(ino) {
-                    to_invalidate.push(entry.parent);
-                    to_invalidate.push(ino);
-                }
-                inodes.remove(ino);
-            }
-            Some(head_info) => {
-                let remote_hash = head_info.xet_hash.as_deref();
-                let remote_etag = head_info.etag.as_deref();
-                // Detect changes via xet_hash (preferred) or ETag.
-                let changed = if current_hash.is_some() || remote_hash.is_some() {
-                    remote_hash != current_hash
-                } else {
-                    remote_etag != current_etag
-                };
-                let mut inodes = self.inode_table.write().expect("inodes poisoned");
-                if changed {
-                    let remote_size = match head_info.size {
-                        Some(s) => s,
-                        None => {
-                            warn!("HEAD response missing size for {}, skipping update", full_path);
-                            return;
-                        }
-                    };
-                    let remote_mtime = head_info
-                        .last_modified
-                        .as_deref()
-                        .map(crate::hub_api::mtime_from_http_date)
-                        .unwrap_or(SystemTime::now());
-                    debug!("Remote change detected via HEAD: {}", full_path);
-                    inodes.update_remote_file(
-                        ino,
-                        remote_hash.map(|s| s.to_string()),
-                        remote_etag.map(|s| s.to_string()),
-                        remote_size,
-                        remote_mtime,
-                    );
-                    to_invalidate.push(ino);
-                } else {
-                    // Update stored etag even when content didn't change,
-                    // so future revalidations have the latest value.
-                    if let Some(entry) = inodes.get_mut(ino)
-                        && remote_etag.is_some()
-                    {
-                        entry.etag = remote_etag.map(|s| s.to_string());
-                    }
-                }
-                // Stamp revalidation time regardless of whether content changed
-                if let Some(entry) = inodes.get_mut(ino) {
-                    entry.last_revalidated = Some(Instant::now());
-                }
-            }
-        }
-
-        // Drop happened at scope end above. Notify after the new state is
-        // published so a racing lookup repopulating the kernel cache cannot
-        // beat the invalidation.
-        if !to_invalidate.is_empty()
-            && let Some(invalidate) = self.invalidator.get()
-        {
-            for ino in to_invalidate {
-                invalidate(ino);
-            }
-        }
-    }
-
-    /// Ensure children of a directory inode are loaded from the Hub API.
-    /// Fetch remote children for `parent_ino` if not already loaded.
-    /// Uses a per-directory lock to prevent thundering herd: concurrent callers
-    /// for the same directory wait on the lock rather than making duplicate HTTP calls.
-    /// Returns ENOENT if the inode doesn't exist, ENOTDIR if it's not a directory.
-    async fn ensure_children_loaded(&self, parent_ino: u64) -> VirtualFsResult<()> {
-        // Fast path: already loaded (no lock needed).
-        {
-            let inodes = self.inode_table.read().expect("inodes poisoned");
-            match inodes.get(parent_ino) {
-                Some(e) if e.kind != InodeKind::Directory => return Err(libc::ENOTDIR),
-                Some(e) if e.children_loaded() => return Ok(()),
-                None => return Err(libc::ENOENT),
-                _ => {}
-            }
-        }
-
-        // Serialize concurrent loads for the same directory.
-        let dir_lock = self.dir_loading_lock(parent_ino);
-        let _guard = dir_lock.lock().await;
-
-        // Re-check: another task may have loaded while we waited.
-        let prefix = {
-            let inodes = self.inode_table.read().expect("inodes poisoned");
-            match inodes.get(parent_ino) {
-                Some(e) if e.kind != InodeKind::Directory => return Err(libc::ENOTDIR),
-                Some(e) if e.children_loaded() => return Ok(()),
-                Some(e) => e.full_path.to_string(),
-                None => return Err(libc::ENOENT),
-            }
-        };
-
-        let entries = match self.hub_client.list_tree(&prefix).await {
-            Ok(entries) => entries,
-            Err(e) => {
-                error!("Failed to list tree for prefix '{}': {}", prefix, e);
-                return Err(libc::EIO);
-            }
-        };
-
-        let mut inodes = self.inode_table.write().expect("inodes poisoned");
-        match inodes.get(parent_ino) {
-            Some(e) if e.children_loaded() => return Ok(()),
-            Some(e) if e.kind != InodeKind::Directory => return Err(libc::ENOTDIR),
-            None => return Err(libc::ENOENT),
-            _ => {}
-        }
-        let mut seen_dirs: std::collections::HashSet<String> = std::collections::HashSet::new();
-        let mut seen_names: std::collections::HashSet<String> = std::collections::HashSet::new();
-
-        for entry in entries {
-            // Convert absolute bucket path to path relative to current directory.
-            // e.g. prefix="models" → "models/bert/config.json" → "bert/config.json"
-            let rel_path = if prefix.is_empty() {
-                entry.path.clone()
-            } else {
-                entry
-                    .path
-                    .strip_prefix(&prefix)
-                    .and_then(|p| p.strip_prefix('/'))
-                    .unwrap_or(&entry.path)
-                    .to_string()
-            };
-
-            if let Some(slash_pos) = rel_path.find('/') {
-                // Nested path → create the immediate subdirectory only (lazy-loaded later)
-                let dir_name = &rel_path[..slash_pos];
-                if seen_dirs.insert(dir_name.to_string()) {
-                    let dir_full_path = if prefix.is_empty() {
-                        dir_name.to_string()
-                    } else {
-                        format!("{}/{}", prefix, dir_name)
-                    };
-                    inodes.insert(
-                        parent_ino,
-                        dir_name.to_string(),
-                        dir_full_path,
-                        InodeKind::Directory,
-                        0,
-                        self.hub_client.default_mtime(),
-                        None,
-                        0o755,
-                        self.uid,
-                        self.gid,
-                    );
-                }
-            } else {
-                let kind = if entry.entry_type == "directory" {
-                    InodeKind::Directory
-                } else {
-                    InodeKind::File
-                };
-                let size = entry.size.unwrap_or(0);
-                let mtime = entry
-                    .mtime
-                    .as_deref()
-                    .map(crate::hub_api::mtime_from_str)
-                    .unwrap_or_else(|| self.hub_client.default_mtime());
-                let default_mode = if kind == InodeKind::Directory { 0o755 } else { 0o644 };
-
-                let ino = inodes.insert(
-                    parent_ino,
-                    rel_path.to_string(),
-                    entry.path,
-                    kind,
-                    size,
-                    mtime,
-                    entry.xet_hash,
-                    default_mode,
-                    self.uid,
-                    self.gid,
-                );
-                let rel_name = rel_path.to_string();
-                seen_names.insert(rel_name);
-                if let Some(oid) = entry.oid
-                    && let Some(e) = inodes.get_mut(ino)
-                {
-                    e.etag = Some(oid);
-                }
-            }
-        }
-
-        // Remove stale children: entries that existed locally but are no longer
-        // in the Hub listing. Skip dirty files (local writes take precedence) and
-        // files with open handles (in-flight reads/writes).
-        if let Some(parent) = inodes.get(parent_ino) {
-            let stale: Vec<u64> = parent
-                .children
-                .iter()
-                .filter(|c| {
-                    let in_listing = seen_names.contains(&*c.name) || seen_dirs.contains(&*c.name);
-                    if in_listing {
-                        return false;
-                    }
-                    inodes.get(c.ino).is_some_and(|e| !e.is_dirty())
-                })
-                .map(|c| c.ino)
-                .collect();
-            for ino in stale {
-                if !inodes.has_dirty_or_open_descendants(ino) {
-                    inodes.remove(ino);
-                }
-            }
-        }
-
-        self.merge_overlay_entries(&mut inodes, parent_ino);
-
-        if let Some(parent) = inodes.get_mut(parent_ino) {
-            // Vec growth doubles capacity; for a one-shot readdir of a stable
-            // directory we'd otherwise keep ~50% slack forever. Trim now,
-            // since regrowth on rare child mutations is cheap.
-            parent.children.shrink_to_fit();
-            parent.children_loaded_at = Some(Instant::now());
-            parent.children_from_remote = true;
-        }
-        Ok(())
-    }
-
-    /// Recursively load all descendants of a directory so in-memory state
-    /// is complete. Used before directory rename to build accurate remote ops.
-    async fn ensure_subtree_loaded(&self, ino: u64) -> VirtualFsResult<()> {
-        self.ensure_children_loaded(ino).await?;
-        // Collect child directories under read lock, then load each recursively
-        let child_dirs: Vec<u64> = {
-            let inodes = self.inode_table.read().expect("inodes poisoned");
-            match inodes.get(ino) {
-                Some(entry) => entry
-                    .children
-                    .iter()
-                    .filter(|c| inodes.get(c.ino).is_some_and(|e| e.kind == InodeKind::Directory))
-                    .map(|c| c.ino)
-                    .collect(),
-                None => return Ok(()),
-            }
-        };
-        for child_dir in child_dirs {
-            Box::pin(self.ensure_subtree_loaded(child_dir)).await?;
-        }
-        Ok(())
-    }
-
-    pub fn alloc_file_handle(&self) -> u64 {
-        self.next_file_handle.fetch_add(1, Ordering::Relaxed)
-    }
-
-    /// Bump the per-inode open-handle refcount. Used by the FUSE adapter
-    /// on `opendir` so a directory with an active readdir can't be evicted.
-    #[cfg(feature = "fuse")]
-    pub(crate) fn bump_open_handles(&self, ino: u64) {
-        self.inode_table.read().expect("inodes poisoned").bump_open_handles(ino);
-    }
-
-    /// Counterpart to `bump_open_handles`, called from `releasedir`.
-    #[cfg(feature = "fuse")]
-    pub(crate) fn drop_open_handles(&self, ino: u64) {
-        self.inode_table.read().expect("inodes poisoned").drop_open_handles(ino);
-    }
-
-    /// Check if any open file handle references the given inode.
-    fn has_open_handles(&self, ino: u64) -> bool {
-        self.inode_table.read().expect("inodes poisoned").has_open_handles(ino)
-    }
-
-    /// Get or create a per-directory lock for serializing ensure_children_loaded().
-    fn dir_loading_lock(&self, ino: u64) -> Arc<tokio::sync::Mutex<()>> {
-        let mut locks = self.dir_loading_locks.lock().expect("dir_loading_locks poisoned");
-        if let Some(weak) = locks.get(&ino)
-            && let Some(arc) = weak.upgrade()
-        {
-            return arc;
-        }
-        locks.retain(|_, w| w.strong_count() > 0);
-        let arc = Arc::new(tokio::sync::Mutex::new(()));
-        locks.insert(ino, Arc::downgrade(&arc));
-        arc
-    }
-
-    fn local_backing_exists(&self, ino: u64, full_path: &str) -> std::io::Result<bool> {
-        match self.overlay_backing.as_deref() {
-            Some(overlay) => overlay.exists(full_path),
-            None => Ok(self
-                .staging
-                .path(ino)
-                .expect("staging directory required for local backing")
-                .exists()),
-        }
-    }
-
-    fn open_local_backing_file(
-        &self,
-        ino: u64,
-        full_path: &str,
-        read: bool,
-        write: bool,
-        create: bool,
-        truncate: bool,
-    ) -> std::io::Result<File> {
-        match self.overlay_backing.as_deref() {
-            Some(overlay) => {
-                // Overlay paths are hierarchical; parents may exist as inodes
-                // (from a remote listing) without being materialized locally.
-                // Staging is flat ino-keyed, so this branch never applies.
-                if create {
-                    overlay.create_parent_dirs(full_path)?;
-                }
-                overlay.open_file(full_path, read, write, create, truncate)
-            }
-            None => {
-                let path = self
-                    .staging
-                    .path(ino)
-                    .expect("staging directory required for local backing");
-                let mut opts = OpenOptions::new();
-                opts.read(read).write(write);
-                if create {
-                    opts.create(true);
-                }
-                if truncate {
-                    opts.truncate(true);
-                }
-                opts.open(&path)
-            }
-        }
-    }
-
-    fn set_local_backing_mode(&self, full_path: &str, mode: u16) -> std::io::Result<()> {
-        match self.overlay_backing.as_deref() {
-            Some(overlay) => overlay.set_mode(full_path, mode),
-            None => Ok(()),
-        }
-    }
-
-    fn remove_local_backing_file(&self, ino: u64, full_path: &str) -> std::io::Result<()> {
-        match self.overlay_backing.as_deref() {
-            Some(overlay) => overlay.remove_file(full_path),
-            None => {
-                let path = self
-                    .staging
-                    .path(ino)
-                    .expect("staging directory required for local backing");
-                std::fs::remove_file(&path)
-            }
-        }
-    }
-
-    /// Install a pending commit watch hook on a streaming channel.
-    /// Called in flush() before commit or deferral so open() can await the result.
-    /// No-op if a hook is already installed (prevents replacing a receiver that
-    /// an open() caller may already be awaiting).
-    fn install_commit_hook(&self, ino: u64, channel: &StreamingChannel) {
-        let mut hook = channel.commit_hook.lock().expect("commit_hook poisoned");
-        if hook.is_some() {
-            return; // already installed — don't replace
-        }
-        let (tx, rx) = tokio::sync::watch::channel(None);
-        *hook = Some(tx);
-        self.pending_commits
-            .lock()
-            .expect("pending_commits poisoned")
-            .insert(ino, rx);
-    }
-
-    /// Fulfill the pending commit hook with a result, then clean up the map.
-    fn fulfill_commit_hook(&self, ino: u64, channel: &StreamingChannel, result: Result<(), i32>) {
-        if let Some(tx) = channel.commit_hook.lock().expect("commit_hook poisoned").take() {
-            let _ = tx.send(Some(result));
-        }
-        self.pending_commits
-            .lock()
-            .expect("pending_commits poisoned")
-            .remove(&ino);
-    }
-
-    /// Wait for any in-flight streaming commit on this inode to complete.
-    /// Returns Ok(()) if no pending commit or commit succeeded, Err(errno) if it failed.
-    async fn await_pending_commit(&self, ino: u64) -> VirtualFsResult<()> {
-        let pending_rx = self
-            .pending_commits
-            .lock()
-            .expect("pending_commits poisoned")
-            .get(&ino)
-            .cloned();
-
-        if let Some(mut rx) = pending_rx {
-            while rx.borrow().is_none() {
-                if rx.changed().await.is_err() {
-                    // Sender dropped without publishing a result — treat as failure.
-                    error!("await_pending_commit: sender dropped for ino={}", ino);
-                    return Err(libc::EIO);
-                }
-            }
-            if let Some(Err(e)) = &*rx.borrow() {
-                debug!("await_pending_commit: commit failed for ino={}: errno={}", ino, e);
-                // Inode was reverted — not an error for the caller, just informational.
-            }
-        }
-        Ok(())
-    }
-
-    /// Set up a new streaming writer + channel. Returns (file_handle, channel).
-    /// Used by both create() and open(O_TRUNC) in simple mode.
-    async fn setup_streaming_writer(
-        &self,
-        pid: Option<u32>,
-        snapshot: InodeSnapshot,
-        dirty_generation_at_open: u64,
-    ) -> VirtualFsResult<(u64, Arc<StreamingChannel>)> {
-        let streaming_writer = self.xet_sessions.create_streaming_writer().await.map_err(|e| {
-            error!("Failed to create streaming writer: {}", e);
-            libc::EIO
-        })?;
-
-        // Bounded channel provides backpressure so a fast writer doesn't queue
-        // unbounded memory. 32 slots × ~128KB FUSE write = ~4MB max in-flight.
-        // blocking_send is safe here: FUSE threads are not tokio workers.
-        let (tx, rx) = tokio::sync::mpsc::channel::<WriteMsg>(32);
-        let error: Arc<std::sync::Mutex<Option<String>>> = Arc::new(std::sync::Mutex::new(None));
-        self.runtime
-            .spawn(streaming_worker(streaming_writer, rx, error.clone()));
-
-        let channel = Arc::new(StreamingChannel {
-            tx,
-            bytes_written: AtomicU64::new(0),
-            error,
-            state: std::sync::Mutex::new(CommitState::Writing),
-            pending_info: std::sync::Mutex::new(None),
-            open_pid: pid,
-            snapshot,
-            dirty_generation_at_open: AtomicU64::new(dirty_generation_at_open),
-            commit_hook: std::sync::Mutex::new(None),
-        });
-
-        let file_handle = self.alloc_file_handle();
-        Ok((file_handle, channel))
-    }
-
-    /// Open a local file as read-only and return the file handle.
-    fn open_local_readonly(&self, ino: u64, path: &PathBuf) -> VirtualFsResult<u64> {
-        match File::open(path) {
-            Ok(file) => self.install_local_handle(ino, Arc::new(Mutex::new(file)), false),
-            Err(e) => {
-                error!("Failed to open file {:?}: {}", path, e);
-                Err(libc::EIO)
-            }
-        }
-    }
-
-    /// Register an already-opened `File` as a read-only `OpenFile::Local`
-    /// handle. Used by the file_cache fast-path so the read fd stays alive
-    /// even if eviction unlinks the on-disk copy after the open.
-    fn install_local_handle(&self, ino: u64, file: Arc<Mutex<File>>, writable: bool) -> VirtualFsResult<u64> {
-        let file_handle = self.alloc_file_handle();
-        {
-            let inodes = self.inode_table.read().expect("inodes poisoned");
-            inodes.bump_open_handles(ino);
-            inodes.touch(ino);
-        }
-        self.open_files
-            .write()
-            .expect("open_files poisoned")
-            .insert(file_handle, OpenFile::Local { ino, file, writable });
-        Ok(file_handle)
-    }
-
-    /// Fire-and-forget background populate of the whole-file cache. Failures
-    /// are logged inside `FileCache::populate`; the read path still works
-    /// via the lazy CAS handle returned to the caller.
-    fn spawn_populate_file_cache(&self, xet_hash: String, file_size: u64) {
-        let Some(fc) = self.file_cache.clone() else { return };
-        let xet = self.xet_sessions.clone();
-        self.runtime.spawn(async move {
-            let hash_for_dl = xet_hash.clone();
-            let _ = fc
-                .populate(&xet_hash, Some(file_size), move |dest| async move {
-                    xet.download_to_file(&hash_for_dl, file_size, &dest).await
-                })
-                .await;
-        });
-    }
-
-    /// Check if a path is in the negative cache (and not expired).
-    fn negative_cache_check(&self, path: &str) -> bool {
-        let cache = self.negative_cache.read().expect("negative_cache poisoned");
-        matches!(cache.get(path), Some(inserted) if inserted.elapsed() < NEG_CACHE_TTL)
-    }
-
-    /// Remove a path from the negative cache (e.g. after create/rename).
-    fn negative_cache_remove(&self, path: &str) {
-        self.negative_cache.write().expect("neg_cache poisoned").remove(path);
-    }
-
-    /// Insert a path into the negative cache, evicting if at capacity.
-    /// Eviction is amortized: instead of scanning all entries, we sample a
-    /// bounded batch to keep the write lock duration constant regardless of cache size.
-    fn negative_cache_insert(&self, path: String) {
-        let mut cache = self.negative_cache.write().expect("neg_cache poisoned");
-        let now = Instant::now();
-        if cache.len() >= NEG_CACHE_CAPACITY {
-            // Evict up to 128 expired entries (bounded scan, not full retain).
-            let expired: Vec<String> = cache
-                .iter()
-                .filter(|(_, ts)| ts.elapsed() >= NEG_CACHE_TTL)
-                .take(128)
-                .map(|(k, _)| k.clone())
-                .collect();
-            for key in &expired {
-                cache.remove(key);
-            }
-            // If still full after evicting expired, drop the oldest sampled entry.
-            if cache.len() >= NEG_CACHE_CAPACITY
-                && let Some(oldest_key) = cache
-                    .iter()
-                    .take(128)
-                    .min_by_key(|(_, ts)| **ts)
-                    .map(|(k, _)| k.clone())
-            {
-                cache.remove(&oldest_key);
-            }
-        }
-        cache.insert(path, now);
-    }
-
-    // ── VFS operations ─────────────────────────────────────────────────
-
-    pub async fn lookup(&self, parent: u64, name: &str) -> VirtualFsResult<VirtualFsAttr> {
-        debug!("lookup: parent={}, name={}", parent, name);
-
-        // Fast path: children already loaded → lookup directly, no allocation needed.
-        // Revalidation info extracted from the lock scope so we can await outside it.
-        enum FastResult {
-            Hit(VirtualFsAttr),
-            NeedsRevalidation {
-                ino: u64,
-                full_path: Arc<str>,
-                current_hash: Option<String>,
-                current_etag: Option<String>,
-            },
-            Miss {
-                full_path: String,
-                /// True when the parent has no remote presence (locally
-                /// created in this session); HEAD/list_tree probes are
-                /// pointless and can be skipped.
-                local_only: bool,
-            },
-            NotLoaded,
-        }
-        let fast = {
-            let inodes = self.inode_table.read().expect("inodes poisoned");
-            let parent_entry = inodes.get(parent).ok_or(libc::ENOENT)?;
-
-            if parent_entry.kind != InodeKind::Directory {
-                return Err(libc::ENOTDIR);
-            }
-
-            // Try in-memory first regardless of `children_loaded`: an inode
-            // inserted via the HEAD point-lookup path or still present after
-            // a partial eviction can be served without a full re-list.
-            //
-            // For directories we still require `children_loaded` to trust the
-            // cached entry: without the parent listing, we can't tell whether
-            // the dir was removed or replaced remotely (files revalidate
-            // individually via HEAD, dirs don't).
-            match inodes.lookup_child(parent, name) {
-                // Clean cached file: HEAD-revalidate (gated by `metadata_ttl`)
-                // so size/hash/mtime stay fresh between poll cycles.
-                Some(entry) if entry.kind == InodeKind::File && !entry.is_dirty() => FastResult::NeedsRevalidation {
-                    ino: entry.inode,
-                    full_path: entry.full_path.clone(),
-                    current_hash: entry.xet_hash.clone(),
-                    current_etag: entry.etag.clone(),
-                },
-                // Either a dirty file (local writes win until flushed) or any
-                // entry under a fully-listed parent we can trust as-is.
-                Some(entry) if entry.kind == InodeKind::File || parent_entry.children_loaded() => {
-                    FastResult::Hit(self.make_vfs_attr(entry))
-                }
-                // Cached non-file under an unloaded parent: we can't HEAD-probe
-                // a directory to check it still exists (resolve endpoint only
-                // serves files), so defer to the slow path which lists.
-                Some(_) => FastResult::NotLoaded,
-                // No cached entry but the parent listing is authoritative →
-                // the name really doesn't exist; populate the negative cache.
-                None if parent_entry.children_loaded() => {
-                    let parent_path = &parent_entry.full_path;
-                    let full_path = if parent_path.is_empty() {
-                        name.to_string()
-                    } else {
-                        format!("{}/{}", parent_path, name)
-                    };
-                    // Skip the HEAD-on-miss revalidation when the listing was
-                    // just refreshed: workloads that rapidly look up unique
-                    // names (tarball extract, build systems, xfstests) would
-                    // otherwise pay one HEAD + list_tree per name despite the
-                    // cache being authoritative.
-                    FastResult::Miss {
-                        full_path,
-                        local_only: !parent_entry.children_from_remote,
-                    }
-                }
-                // No entry, no listing → slow path (HEAD then list_tree).
-                None => FastResult::NotLoaded,
-            }
-        }; // inodes lock dropped here
-
-        match fast {
-            FastResult::Hit(attr) => return Ok(attr),
-            FastResult::NeedsRevalidation {
-                ino,
-                full_path,
-                current_hash,
-                current_etag,
-            } => {
-                self.revalidate_file(ino, &full_path, current_hash.as_deref(), current_etag.as_deref())
-                    .await;
-                let inodes = self.inode_table.read().expect("inodes poisoned");
-                return match inodes.get(ino) {
-                    Some(entry) => Ok(self.make_vfs_attr(entry)),
-                    None => Err(libc::ENOENT),
-                };
-            }
-            FastResult::Miss { full_path, local_only } => {
-                // A cached listing can still hide entries added remotely after
-                // we listed. Re-validate before serving ENOENT, gated by the
-                // negative cache so repeated misses stay free.
-                if self.negative_cache_check(&full_path) {
-                    return Err(libc::ENOENT);
-                }
-                // Skip HEAD/list probes for purely-local directories
-                // (locally-mkdir'd this session, never seen on remote):
-                // there's no remote state to discover, so the cached listing
-                // is authoritative. Hot path for tarball extract / xfstests
-                // creating thousands of unique names under fresh dirs.
-                //
-                // Do NOT seed the global negative cache here — its TTL is
-                // longer than metadata_ttl, and remote churn could materialize
-                // these names while the entry hides them. The listing itself
-                // is the cache for these misses.
-                if local_only {
-                    return Err(libc::ENOENT);
-                }
-                // Overlay precedence: a file added to the overlay backing
-                // after the parent listing was loaded must shadow whatever the
-                // remote HEAD returns. Otherwise a clean remote inode wins,
-                // and a later open(write) would truncate the local backing.
-                if self.overlay_lookup_takes_precedence(parent, &full_path)? {
-                    let inodes = self.inode_table.read().expect("inodes poisoned");
-                    return inodes
-                        .lookup_child(parent, name)
-                        .map(|e| self.make_vfs_attr(e))
-                        .ok_or(libc::ENOENT);
-                }
-                // HEAD probe catches files added remotely.
-                if let Ok(Some(head)) = self.hub_client.head_file(&full_path).await
-                    && head.size.is_some()
-                {
-                    return self.insert_file_from_head(parent, name, &full_path, head);
-                }
-                // The resolve endpoint returns 404 for directories, so a HEAD
-                // miss could still be a remotely-added dir. Targeted listing
-                // catches that; non-empty result means the dir exists.
-                if let Ok(entries) = self.hub_client.list_tree(&full_path).await
-                    && !entries.is_empty()
-                {
-                    return self.insert_dir(parent, name, &full_path);
-                }
-                self.negative_cache_insert(full_path);
-                return Err(libc::ENOENT);
-            }
-            FastResult::NotLoaded => {} // fall through to slow path
-        }
-
-        // Slow path: children not loaded yet, fetch from Hub API.
-        let full_path = {
-            let inodes = self.inode_table.read().expect("inodes poisoned");
-            match inodes.get(parent) {
-                Some(e) => inode::child_path(&e.full_path, name),
-                None => return Err(libc::ENOENT),
-            }
-        };
-
-        if self.negative_cache_check(&full_path) {
-            debug!("negative cache hit: {}", full_path);
-            return Err(libc::ENOENT);
-        }
-
-        // Overlay precedence (see FastResult::Miss): defer to the parent
-        // listing + merge so the local entry is what we resolve.
-        if self.overlay_lookup_takes_precedence(parent, &full_path)? {
-            self.ensure_children_loaded(parent).await?;
-            let inodes = self.inode_table.read().expect("inodes poisoned");
-            return inodes
-                .lookup_child(parent, name)
-                .map(|e| self.make_vfs_attr(e))
-                .ok_or(libc::ENOENT);
-        }
-
-        // Resolve the single requested path via HEAD instead of listing the
-        // whole parent — for point-access workloads (no `readdir`) this keeps
-        // the inode table scoped to what the caller actually touches.
-        match self.hub_client.head_file(&full_path).await {
-            // Without size, `open_readonly` would take the empty-file shortcut
-            // and expose a zero-byte file. Fall back to list_tree which has
-            // the authoritative size from the tree index.
-            Ok(Some(head)) if head.size.is_some() => {
-                return self.insert_file_from_head(parent, name, &full_path, head);
-            }
-            // 404 may mean "doesn't exist" or "it's a directory" (the resolve
-            // endpoint only handles files), so the listing has the final word.
-            Ok(_) => {}
-            Err(e) => debug!("HEAD lookup {} failed, falling back to list: {}", full_path, e),
-        }
-
-        self.ensure_children_loaded(parent).await?;
-
-        let inodes = self.inode_table.read().expect("inodes poisoned");
-        match inodes.lookup_child(parent, name) {
-            Some(entry) => Ok(self.make_vfs_attr(entry)),
-            None => {
-                drop(inodes);
-                self.negative_cache_insert(full_path);
-                Err(libc::ENOENT)
-            }
-        }
-    }
-
-    /// Insert a file inode resolved via HEAD, without touching its siblings.
-    /// Used by the lookup slow path to avoid listing the whole parent dir
-    /// when only one file is needed. The parent's `children_loaded` stays
-    /// `false` so a later `readdir` still triggers a full listing.
-    fn insert_file_from_head(
-        &self,
-        parent: u64,
-        name: &str,
-        full_path: &str,
-        head: crate::hub_api::HeadFileInfo,
-    ) -> VirtualFsResult<VirtualFsAttr> {
-        let size = head.size.unwrap_or(0);
-        let mtime = head
-            .last_modified
-            .as_deref()
-            .map(crate::hub_api::mtime_from_http_date)
-            .unwrap_or_else(|| self.hub_client.default_mtime());
-
-        let mut inodes = self.inode_table.write().expect("inodes poisoned");
-        match inodes.get(parent) {
-            None => return Err(libc::ENOENT),
-            Some(e) if e.kind != InodeKind::Directory => return Err(libc::ENOTDIR),
-            Some(_) => {}
-        }
-        if let Some(existing) = inodes.lookup_child(parent, name) {
-            // A concurrent HEAD lookup may have just inserted the same file —
-            // return its attr instead of redoing the work.
-            if existing.kind == InodeKind::File {
-                return Ok(self.make_vfs_attr(existing));
-            }
-            // The cached entry is a directory but HEAD just confirmed the
-            // remote path is a file: the dir was replaced. Evict the stale
-            // subtree, unless dirty/open descendants would be lost — in that
-            // case keep serving the stale dir (the next sweep / explicit
-            // close will eventually clear the way).
-            let stale_ino = existing.inode;
-            if inodes.has_dirty_or_open_descendants(stale_ino) {
-                return Ok(self.make_vfs_attr(existing));
-            }
-            inodes.remove(stale_ino);
-        }
-        let ino = inodes.insert(
-            parent,
-            name.to_string(),
-            full_path.to_string(),
-            InodeKind::File,
-            size,
-            mtime,
-            head.xet_hash,
-            0o644,
-            self.uid,
-            self.gid,
-        );
-        if let Some(etag) = head.etag
-            && let Some(entry) = inodes.get_mut(ino)
-        {
-            entry.etag = Some(etag);
-        }
-        match inodes.get(ino) {
-            Some(entry) => Ok(self.make_vfs_attr(entry)),
-            None => Err(libc::EIO),
-        }
-    }
-
-    /// Insert a directory inode discovered via a targeted listing, without
-    /// touching siblings. Mirrors `insert_file_from_head` for the dir case.
-    /// Parent's `children_loaded` stays as-is so a later `readdir` still
-    /// triggers a full listing.
-    fn insert_dir(&self, parent: u64, name: &str, full_path: &str) -> VirtualFsResult<VirtualFsAttr> {
-        let mut inodes = self.inode_table.write().expect("inodes poisoned");
-        match inodes.get(parent) {
-            None => return Err(libc::ENOENT),
-            Some(e) if e.kind != InodeKind::Directory => return Err(libc::ENOTDIR),
-            Some(_) => {}
-        }
-        if let Some(existing) = inodes.lookup_child(parent, name) {
-            return Ok(self.make_vfs_attr(existing));
-        }
-        let ino = inodes.insert(
-            parent,
-            name.to_string(),
-            full_path.to_string(),
-            InodeKind::Directory,
-            0,
-            self.hub_client.default_mtime(),
-            None,
-            0o755,
-            self.uid,
-            self.gid,
-        );
-        match inodes.get(ino) {
-            Some(entry) => Ok(self.make_vfs_attr(entry)),
-            None => Err(libc::EIO),
-        }
-    }
-
-    pub fn default_uid(&self) -> u32 {
-        self.uid
-    }
-
-    pub fn default_gid(&self) -> u32 {
-        self.gid
-    }
-
-    /// Schedule a debounced flush for a dirty inode.
-    /// Used by NFS (which has no close/flush RPC) to ensure writes
-    /// eventually get committed to the Hub.
-    pub fn schedule_flush(&self, ino: u64) {
-        if let Some(fm) = &self.flush_manager {
-            fm.enqueue(ino);
-        }
-    }
-
-    /// Enqueues the inode for background flush rather than blocking on a
-    /// synchronous Hub upload. Data is already durable in the local staging
-    /// file after write(); the background flush loop commits to Hub.
-    ///
-    /// - Streaming handles: no-op (committed synchronously on close).
-    /// - Advanced-writes handles: enqueues for background flush.
-    /// - Read-only / lazy handles: no-op.
-    pub async fn fsync(&self, ino: u64, file_handle: u64, _pid: Option<u32>) -> VirtualFsResult<()> {
-        let overlay_file = {
-            let files = self.open_files.read().expect("open_files poisoned");
-            match files.get(&file_handle) {
-                Some(OpenFile::Streaming { .. }) => return Ok(()),
-                // Overlay: sync local fd for durability, skip remote upload.
-                Some(OpenFile::Local { file, .. }) if self.overlay() => Some(Arc::clone(file)),
-                _ => None,
-            }
-        };
-        if let Some(file) = overlay_file {
-            let file = file.lock().expect("local file mutex poisoned");
-            return file.sync_all().map_err(|e| {
-                error!("Overlay fsync failed for ino={}: {}", ino, e);
-                libc::EIO
-            });
-        }
-        self.schedule_flush(ino);
-        Ok(())
-    }
-
-    pub fn getattr(&self, ino: u64) -> VirtualFsResult<VirtualFsAttr> {
-        debug!("getattr: ino={}", ino);
-
-        let inodes = self.inode_table.read().expect("inodes poisoned");
-        match inodes.get(ino) {
-            Some(entry) => {
-                inodes.touch(ino);
-                Ok(self.make_vfs_attr(entry))
-            }
-            None => Err(libc::ENOENT),
-        }
-    }
-
-    /// Must be called after every `reply.entry()` / `reply.created()`, or
-    /// the kernel and our `nlookup` will drift and a later `forget()` will
-    /// underflow. Also touches for LRU so actively-looked-up inodes aren't
-    /// immediately evicted.
-    #[cfg(any(feature = "fuse", test))]
-    pub(crate) fn bump_nlookup(&self, ino: u64) {
-        let inodes = self.inode_table.read().expect("inodes poisoned");
-        inodes.bump_nlookup(ino);
-        inodes.touch(ino);
-    }
-
-    /// Handle a FUSE `forget(ino, nlookup)`: drop the refcount by `nlookup`,
-    /// and evict the inode if it's now safe (kernel no longer holds the
-    /// dentry, no open handles, nothing dirty). Eviction keeps `InodeTable`
-    /// bounded — without it the table grows for every file ever looked up.
-    #[cfg(any(feature = "fuse", test))]
-    pub(crate) fn forget(&self, ino: u64, nlookup: u64) {
-        debug!("forget: ino={} nlookup={}", ino, nlookup);
-
-        // Shared lock for the hot path: dropping the refcount is an atomic op,
-        // so readers (lookup/getattr/read) aren't blocked on the common case
-        // where the refcount stays > 0.
-        let reached_zero = {
-            let inodes = self.inode_table.read().expect("inodes poisoned");
-            inodes.drop_nlookup(ino, nlookup)
-        };
-        if !reached_zero {
-            return;
-        }
-
-        // A racing open() would insert into `open_files` before calling
-        // `bump_nlookup`, so checking handles here (after the refcount hit 0)
-        // is safe: any concurrent lookup will re-bump the count, and any
-        // concurrent read/write holds an open handle we'll see.
-        if self.has_open_handles(ino) {
-            // The kernel has given up the dentry but our handle keeps the
-            // inode alive. Mark it so `release()` finishes the eviction
-            // once the last handle closes — otherwise it would leak.
-            self.inode_table
-                .read()
-                .expect("inodes poisoned")
-                .mark_evict_pending(ino);
-            return;
-        }
-
-        let evicted = {
-            let mut inodes = self.inode_table.write().expect("inodes poisoned");
-            inodes.evict_if_safe(ino)
-        };
-        if evicted {
-            self.drop_staging(ino);
-        }
-    }
-
-    /// Remove any on-disk staging file for `ino`. Safe to call for inodes that
-    /// never had a staging file — NotFound is ignored. Keeps `StagingDir`'s
-    /// byte budget accurate via `try_remove`.
-    fn drop_staging(&self, ino: u64) {
-        if let Some(sd) = self.staging.dir() {
-            sd.try_remove(ino);
-        }
-    }
-
-    pub async fn readdir(&self, ino: u64) -> VirtualFsResult<Vec<VirtualFsDirEntry>> {
-        debug!("readdir: ino={}", ino);
-
-        self.ensure_children_loaded(ino).await?;
-
-        let inodes = self.inode_table.read().expect("inodes poisoned");
-        let entry = inodes.get(ino).ok_or(libc::ENOENT)?;
-
-        let mut entries = Vec::with_capacity(2 + entry.children.len());
-        entries.push(VirtualFsDirEntry {
-            ino,
-            kind: InodeKind::Directory,
-            name: ".".to_string(),
-        });
-        entries.push(VirtualFsDirEntry {
-            ino: entry.parent,
-            kind: InodeKind::Directory,
-            name: "..".to_string(),
-        });
-
-        for child_ref in &entry.children {
-            if let Some(child) = inodes.get(child_ref.ino) {
-                entries.push(VirtualFsDirEntry {
-                    ino: child.inode,
-                    kind: child.kind,
-                    name: child_ref.name.to_string(),
-                });
-            }
-        }
-
-        Ok(entries)
-    }
-
-    pub async fn open(&self, ino: u64, writable: bool, truncate: bool, pid: Option<u32>) -> VirtualFsResult<u64> {
-        debug!(
-            "open: ino={}, writable={}, truncate={}, pid={:?}",
-            ino, writable, truncate, pid
-        );
-
-        if writable && self.read_only {
-            return Err(libc::EROFS);
-        }
-
-        let file_entry = self.get_file_entry(ino)?;
-        let staging_path = self.staging.path(ino);
-
-        if writable && self.advanced_writes {
-            // Staging file + async flush (supports random writes and seek)
-            self.open_advanced_write(
-                ino,
-                &file_entry.full_path,
-                &file_entry.xet_hash,
-                file_entry.size,
-                truncate,
-            )
-            .await
-        } else if writable && truncate {
-            // Simple streaming write (append-only, synchronous commit on close)
-            self.open_streaming_write(ino, pid).await
-        } else if writable {
-            // Simple mode without O_TRUNC: random writes not supported
-            Err(libc::EPERM)
-        } else {
-            self.open_readonly(ino, file_entry, staging_path).await
-        }
-    }
-
-    /// Advanced writes: prepare a staging file and open it for read-write.
-    async fn open_advanced_write(
-        &self,
-        ino: u64,
-        full_path: &str,
-        xet_hash: &str,
-        size: u64,
-        truncate: bool,
-    ) -> VirtualFsResult<u64> {
-        // Serialize staging preparation per inode (prevents concurrent download races)
-        let staging_mutex = self.staging.lock(ino);
-        let _staging_guard = staging_mutex.lock().await;
-
-        // Reuse the staging file when either (a) it has pending dirty writes,
-        // or (b) it's a clean cache flagged as current. In both cases its
-        // content is the right starting point for this open.
-        let (is_dirty, staging_is_current) = {
-            let inodes = self.inode_table.read().expect("inodes poisoned");
-            let entry = inodes.get(ino).ok_or(libc::ENOENT)?;
-            (entry.is_dirty(), entry.staging_is_current)
-        };
-        let local_exists = self.local_backing_exists(ino, full_path).map_err(|e| {
-            error!("Failed to check local backing file for ino={}: {}", ino, e);
-            libc::EIO
-        })?;
-        // Overlay never downloads from remote — only existing local files (or dirty drafts) are writable.
-        if self.overlay() && !is_dirty && !local_exists {
-            return Err(libc::EPERM);
-        }
-
-        let can_reuse_staging = !truncate && (is_dirty || staging_is_current) && local_exists;
-
-        if !can_reuse_staging {
-            // Clear the flag before touching disk so a partial failure (e.g.
-            // mid-download CAS error, async cancel) never leaves the cache
-            // reusable.
-            if let Some(entry) = self.inode_table.write().expect("inodes poisoned").get_mut(ino) {
-                entry.staging_is_current = false;
-            }
-            // GC accounting only matters for non-overlay (overlay files live
-            // in user dir, so file_size returns 0 here on miss).
-            let old_size = self.staging.dir().map(|sd| sd.file_size(ino)).unwrap_or(0);
-            let needs_download = !self.overlay() && !truncate && !xet_hash.is_empty() && size > 0;
-            let new_size = if needs_download {
-                let staging_path = self
-                    .staging
-                    .path(ino)
-                    .expect("staging directory required for advanced writes");
-                self.xet_sessions
-                    .download_to_file(xet_hash, size, &staging_path)
-                    .await
-                    .map_err(|e| {
-                        error!("Failed to download file for write: {}", e);
-                        libc::EIO
-                    })?;
-                size
-            } else {
-                self.open_local_backing_file(ino, full_path, true, true, true, true)
-                    .map_err(|e| {
-                        error!("Failed to create local backing file: {}", e);
-                        libc::EIO
-                    })?;
-                0
-            };
-            if !self.overlay()
-                && let Some(sd) = self.staging.dir()
-            {
-                sd.resize_bytes(old_size, new_size);
-            }
-            // Flag the cache as current only when the staging actually mirrors
-            // the remote. Cases to exclude:
-            // - truncated hashed file: empty staging, non-empty xet_hash.
-            // - non-Xet file with `size > 0` and no xet_hash: File::create
-            //   leaves empty staging, which does not match the remote.
-            // - race with poll: `xet_hash` moved between `open()` reading the
-            //   inode and here, so the downloaded hash is now stale — detected
-            //   by the `entry.xet_hash == xet_hash` post-check.
-            // Skip in overlay mode: there's no remote materialization concept.
-            let materializes_remote = !self.overlay() && (needs_download || (xet_hash.is_empty() && size == 0));
-            if materializes_remote
-                && let Some(entry) = self.inode_table.write().expect("inodes poisoned").get_mut(ino)
-                && entry.xet_hash.as_deref().unwrap_or("") == xet_hash
-            {
-                entry.staging_is_current = true;
-            }
-        }
-        let file = self
-            .open_local_backing_file(ino, full_path, true, true, false, false)
-            .map_err(|e| {
-                error!("Failed to open staging file: {}", e);
-                libc::EIO
-            })?;
-
-        // Re-check inode still exists before committing the open
-        {
-            let mut inodes = self.inode_table.write().expect("inodes poisoned");
-            let entry = inodes.get_mut(ino).ok_or(libc::ENOENT)?;
-            entry.set_dirty();
-            if truncate {
-                entry.size = 0;
-                // POSIX: O_TRUNC must update mtime and ctime
-                let now = SystemTime::now();
-                entry.mtime = now;
-                entry.ctime = now;
-            }
-        }
-
-        self.install_local_handle(ino, Arc::new(Mutex::new(file)), true)
-    }
-
-    /// Simple streaming write: truncate existing file and set up a new streaming writer.
-    async fn open_streaming_write(&self, ino: u64, pid: Option<u32>) -> VirtualFsResult<u64> {
-        // Wait for any in-flight commit to complete before starting a new writer.
-        self.await_pending_commit(ino).await?;
-
-        let staging_mutex = self.staging.lock(ino);
-        let _staging_guard = staging_mutex.lock().await;
-
-        // Capture inode snapshot before mutation (for revert on commit failure)
-        let snapshot = {
-            let inodes = self.inode_table.read().expect("inodes poisoned");
-            let entry = inodes.get(ino).ok_or(libc::ENOENT)?;
-            InodeSnapshot {
-                xet_hash: entry.xet_hash.clone(),
-                size: entry.size,
-                mtime: entry.mtime,
-                pending_deletes: entry.pending_deletes.clone(),
-                existed_before: true,
-            }
-        };
-
-        let (file_handle, channel) = self.setup_streaming_writer(pid, snapshot, 0).await?;
-
-        {
-            let mut inodes = self.inode_table.write().expect("inodes poisoned");
-            if let Some(entry) = inodes.get_mut(ino) {
-                entry.set_dirty();
-                entry.size = 0;
-                entry.xet_hash = None;
-                channel
-                    .dirty_generation_at_open
-                    .store(entry.dirty_generation, Ordering::Relaxed);
-            }
-        }
-
-        self.inode_table.read().expect("inodes poisoned").bump_open_handles(ino);
-        self.open_files
-            .write()
-            .expect("open_files poisoned")
-            .insert(file_handle, OpenFile::Streaming { ino, channel });
-        Ok(file_handle)
-    }
-
-    /// Open a file for reading. Dispatches based on where the content lives.
-    async fn open_readonly(&self, ino: u64, fe: FileEntry, staging_path: Option<PathBuf>) -> VirtualFsResult<u64> {
-        // For dirty-staging reads, hold the per-inode staging lock across
-        // both the existence check and the open so the flush-path GC can't
-        // unlink the file in between (it takes the same lock in gc_one).
-        let _staging_guard = if fe.is_dirty && staging_path.is_some() {
-            Some(self.staging.lock(ino).lock_owned().await)
-        } else {
-            None
-        };
-
-        // Overlay mode: dirty file is in the overlay backing, not the staging path.
-        if fe.is_dirty && self.overlay() {
-            let local_exists = self.local_backing_exists(ino, &fe.full_path).map_err(|e| {
-                error!("Failed to check local backing file for {}: {}", fe.full_path, e);
-                libc::EIO
-            })?;
-            if local_exists {
-                let file = self
-                    .open_local_backing_file(ino, &fe.full_path, true, false, false, false)
-                    .map_err(|e| {
-                        error!("Failed to open local backing file {:?}: {}", fe.full_path, e);
-                        libc::EIO
-                    })?;
-                return self.install_local_handle(ino, Arc::new(Mutex::new(file)), false);
-            }
-            error!("Dirty overlay file ino={} has missing local backing file", ino);
-            return Err(libc::EIO);
-        }
-
-        // If the dirty snapshot pointed at a now-missing staging file, the
-        // flush-path GC may have raced before we took the lock above: the
-        // inode was flushed clean and its staging reclaimed between the
-        // snapshot and here. Re-read the entry so we dispatch on current
-        // state instead of returning EIO.
-        let fe = match (fe.is_dirty, &staging_path) {
-            (true, Some(p)) if !p.exists() => self.get_file_entry(ino)?,
-            _ => fe,
-        };
-        match (fe.is_dirty, &staging_path) {
-            // Advanced write in progress — read from local staging file.
-            (true, Some(path)) if path.exists() => self.open_local_readonly(ino, path),
-
-            // Dirty file but staging file is missing — should not happen.
-            (true, Some(_)) => {
-                error!("Dirty file ino={} has missing staging file", ino);
-                Err(libc::EIO)
-            }
-
-            // Dirty without staging path — committed since the snapshot read.
-            // Re-fetch entry and dispatch on the now-clean state.
-            (true, None) => {
-                let fe = self.get_file_entry(ino)?;
-                if !fe.xet_hash.is_empty() {
-                    return self.open_lazy(ino, fe.xet_hash, fe.size);
-                }
-                self.open_lazy(ino, String::new(), 0)
-            }
-
-            // Remote xet-backed file — try the whole-file cache first, then
-            // fall back to lazy CAS range reads.
-            _ if !fe.xet_hash.is_empty() => {
-                if let Some(fc) = &self.file_cache {
-                    if let Some(file) = fc.try_open(&fe.xet_hash).await {
-                        let cloned = std::fs::File::try_clone(&*file).map_err(|_| libc::EIO)?;
-                        return self.install_local_handle(ino, Arc::new(Mutex::new(cloned)), false);
-                    }
-                    self.spawn_populate_file_cache(fe.xet_hash.clone(), fe.size);
-                }
-                self.open_lazy(ino, fe.xet_hash, fe.size)
-            }
-
-            // Non-Xet file — HTTP download to staging cache.
-            //
-            // TODO(staging-gc): http_<hash> files are not counted in
-            // StagingDir::bytes_used and not picked up by try_remove(ino), so
-            // a repo with large non-Xet files can exceed --max-staging-size
-            // without the GC noticing. Either resize_bytes() after download +
-            // index http_* paths in StagingCoordinator, or document that the
-            // budget covers writes only.
-            _ if fe.size > 0 => {
-                let staging = self.staging.dir().ok_or_else(|| {
-                    error!("No staging dir for HTTP download of ino={}", ino);
-                    libc::EIO
-                })?;
-                let path_hash = {
-                    use std::hash::{Hash, Hasher};
-                    let mut h = std::collections::hash_map::DefaultHasher::new();
-                    self.hub_client.source().to_string().hash(&mut h);
-                    fe.full_path.hash(&mut h);
-                    h.finish()
-                };
-                let dest = staging.root().join(format!("http_{:x}", path_hash));
-                {
-                    let lock = self.staging.lock(ino);
-                    let _guard = lock.lock().await;
-                    // download_file_http uses ETag-based conditional requests,
-                    // so this is cheap when the cached file is still valid (304).
-                    self.hub_client
-                        .download_file_http(&fe.full_path, &dest)
-                        .await
-                        .map_err(|e| {
-                            error!("HTTP download failed for {}: {}", fe.full_path, e);
-                            libc::EIO
-                        })?;
-                }
-                self.open_local_readonly(ino, &dest)
-            }
-
-            // Empty file (size=0, no hash).
-            _ => self.open_lazy(ino, String::new(), 0),
-        }
-    }
-
-    /// Allocate a lazy file handle backed by a prefetch buffer.
-    fn open_lazy(&self, ino: u64, xet_hash: String, size: u64) -> VirtualFsResult<u64> {
-        let prefetch = Arc::new(tokio::sync::Mutex::new(PrefetchState::new(
-            xet_hash,
-            size,
-            self.direct_io,
-        )));
-        let file_handle = self.alloc_file_handle();
-        self.inode_table.read().expect("inodes poisoned").bump_open_handles(ino);
-        self.open_files
-            .write()
-            .expect("open_files poisoned")
-            .insert(file_handle, OpenFile::Lazy { ino, prefetch });
-        Ok(file_handle)
-    }
-
-    /// Fetch data for the prefetch buffer. Uses the persistent stream for sequential
-    /// reads (with automatic (re)start), or opens a temporary stream with retries
-    /// for range/seek access. Returns EIO if all attempts fail.
-    async fn fetch_data(
-        &self,
-        prefetch_state: &mut PrefetchState,
-        cursor: u64,
-        plan: &FetchPlan,
-        file_size: u64,
-    ) -> std::result::Result<(VecDeque<Bytes>, usize), i32> {
-        const MAX_ATTEMPTS: u32 = 3;
-        let file_info = XetFileInfo::new(prefetch_state.xet_hash.clone(), prefetch_state.file_size);
-
-        // Take the persistent stream before the loop (sequential reads).
-        // first_stream.take() only returns Some on the first iteration.
-        let mut first_stream = if plan.strategy.is_stream() {
-            prefetch_state.stream.take()
-        } else {
-            None
-        };
-
-        for attempt in 0..MAX_ATTEMPTS {
-            let stream = match first_stream.take() {
-                Some(s) => Some(s),
-                None => match self.xet_sessions.download_stream_boxed(
-                    &file_info,
-                    cursor,
-                    // For range downloads (random reads / seeks), bound the request
-                    // so CAS only returns terms covering the needed bytes.
-                    // For streaming, leave unbounded so the stream can continue.
-                    if plan.strategy == prefetch::FetchStrategy::RangeDownload {
-                        Some(cursor + plan.fetch_size)
-                    } else {
-                        None
-                    },
-                ) {
-                    Ok(s) => Some(s),
-                    Err(e) => {
-                        warn!(
-                            "prefetch: stream open failed: cursor={}, attempt={}/{}: {}",
-                            cursor,
-                            attempt + 1,
-                            MAX_ATTEMPTS,
-                            e,
-                        );
-                        None
-                    }
-                },
-            };
-
-            if let Some(mut stream) = stream {
-                // Read chunks until fetch_size bytes
-                let mut chunks = VecDeque::new();
-                let mut total = 0usize;
-                let mut failed = false;
-                let mut stream_eof = false;
-                while (total as u64) < plan.fetch_size {
-                    match stream.next().await {
-                        Ok(Some(chunk)) => {
-                            total += chunk.len();
-                            chunks.push_back(chunk);
-                        }
-                        Ok(None) => {
-                            stream_eof = true;
-                            break;
-                        }
-                        Err(e) => {
-                            warn!(
-                                "prefetch: stream read error at cursor={}, got={}/{}, attempt={}/{}: {}",
-                                cursor,
-                                total,
-                                plan.fetch_size,
-                                attempt + 1,
-                                MAX_ATTEMPTS,
-                                e,
-                            );
-                            failed = true;
-                            break;
-                        }
-                    }
-                }
-
-                // Early EOF sanity check: only meaningful when we got some data
-                // (zero-data EOF is just a failed attempt, handled by retry below)
-                if stream_eof && total > 0 {
-                    let stream_total = cursor + total as u64;
-                    if stream_total < file_size {
-                        error!(
-                            "prefetch: stream EOF at cursor={} after {} bytes (file_size={}, \
-                             shortfall={}, hash={})",
-                            cursor,
-                            total,
-                            file_size,
-                            file_size - stream_total,
-                            prefetch_state.xet_hash,
-                        );
-                        debug_assert!(
-                            false,
-                            "stream EOF before file_size: got {stream_total}, expected {file_size}"
-                        );
-                    }
-                } else if plan.strategy.is_stream() && !failed {
-                    // Keep stream alive for future sequential reads
-                    prefetch_state.stream = Some(stream);
-                }
-
-                if total > 0 {
-                    debug!(
-                        "prefetch fetch: cursor={}, got={}, window={}",
-                        cursor, total, prefetch_state.window_size,
-                    );
-                    return Ok((chunks, total));
-                }
-            }
-
-            if attempt + 1 < MAX_ATTEMPTS {
-                tokio::time::sleep(Duration::from_millis(100 * (attempt as u64 + 1))).await;
-            }
-        }
-
-        error!(
-            "prefetch: all {} fetch attempts failed: cursor={}, hash={}",
-            MAX_ATTEMPTS, cursor, prefetch_state.xet_hash,
-        );
-        Err(libc::EIO)
-    }
-
-    /// Read data from an open file. Returns `(data, eof)`.
-    pub async fn read(&self, file_handle: u64, offset: u64, size: u32) -> VirtualFsResult<(Bytes, bool)> {
-        debug!("read: fh={}, offset={}, size={}", file_handle, offset, size);
-
-        // Extract what we need under the lock, then release it.
-        // Clone Arc<File> so the FD stays alive even if release() runs concurrently.
-        let read_target = {
-            let files = self.open_files.read().expect("open_files poisoned");
-            match files.get(&file_handle) {
-                Some(OpenFile::Local { file, .. }) => ReadTarget::LocalFd(file.clone()),
-                Some(OpenFile::Lazy { prefetch, .. }) => ReadTarget::Remote {
-                    prefetch: prefetch.clone(),
-                },
-                // write-only, not readable
-                Some(OpenFile::Streaming { .. }) => return Err(libc::EBADF), // write-only, not readable
-                None => return Err(libc::EBADF), // handle already closed (race with release)
-            }
-        };
-
-        match read_target {
-            ReadTarget::LocalFd(file) => {
-                let mut file = file.lock().expect("local file mutex poisoned");
-                let mut buf = BytesMut::zeroed(size as usize);
-                file.seek(SeekFrom::Start(offset)).map_err(|_| libc::EIO)?;
-                let n = file.read(&mut buf).map_err(|_| libc::EIO)?;
-                buf.truncate(n);
-                let eof = (n as u32) < size;
-                Ok((buf.freeze(), eof))
-            }
-            ReadTarget::Remote { prefetch } => {
-                let mut prefetch_state = prefetch.lock().await;
-
-                let file_size = prefetch_state.file_size;
-
-                // Past EOF
-                if offset >= file_size {
-                    return Ok((Bytes::new(), true));
-                }
-
-                // Maximum bytes we should return (capped at file boundary).
-                let to_read = ((size as u64).min(file_size - offset)) as usize;
-
-                // IMPORTANT: Never return a short read unless at real EOF.
-                // The Linux FUSE kernel module shrinks i_size on short reads
-                // (fuse_read_update_size), which makes subsequent reads return
-                // 0 bytes and truncates the file from the application's view.
-
-                let mut response = BytesMut::with_capacity(to_read);
-                let mut cursor = offset;
-
-                // Fast path: forward buffer has enough data (common case, zero-copy).
-                if let Some(data) = prefetch_state.try_serve_forward(offset, size) {
-                    if data.len() == to_read {
-                        debug!("prefetch hit (forward): offset={}, len={}", offset, data.len());
-                        let eof = offset + data.len() as u64 >= file_size;
-                        return Ok((data, eof));
-                    }
-                    cursor += data.len() as u64;
-                    response.extend_from_slice(&data);
-                } else if let Some(data) = prefetch_state.try_serve_seek(offset, size) {
-                    // Seek window: only reachable when forward buffer has no data
-                    // at this offset (backward seek). Zero-copy on full hit.
-                    if data.len() == to_read {
-                        debug!("prefetch hit (seek): offset={}, len={}", offset, data.len());
-                        let eof = offset + data.len() as u64 >= file_size;
-                        return Ok((data, eof));
-                    }
-                    cursor += data.len() as u64;
-                    response.extend_from_slice(&data);
-                }
-
-                // Assembly loop: fetch more data until we have to_read bytes.
-                // Only reached at prefetch window boundaries or cache misses.
-                while response.len() < to_read {
-                    let remaining = (to_read - response.len()) as u32;
-                    let plan = prefetch_state.prepare_fetch(cursor, remaining);
-                    debug!(
-                        "prefetch miss: cursor={}, remaining={}, strategy={:?}, fetch_size={}, \
-                         buf_start={}, file_size={}, has_stream={}",
-                        cursor,
-                        remaining,
-                        plan.strategy,
-                        plan.fetch_size,
-                        prefetch_state.buf_start,
-                        file_size,
-                        prefetch_state.stream.is_some(),
-                    );
-
-                    let (chunks, total) = self.fetch_data(&mut prefetch_state, cursor, &plan, file_size).await?;
-
-                    prefetch_state.store_fetched(cursor, chunks, total);
-
-                    // Drain freshly filled buffer into response
-                    let remaining = (to_read - response.len()) as u32;
-                    if let Some(data) = prefetch_state.try_serve_forward(cursor, remaining) {
-                        cursor += data.len() as u64;
-                        response.extend_from_slice(&data);
-                    }
-                }
-
-                let eof = cursor == file_size;
-                Ok((response.freeze(), eof))
-            }
-        }
-    }
-
-    pub fn write(&self, ino: u64, file_handle: u64, offset: u64, data: &[u8]) -> VirtualFsResult<u32> {
-        debug!(
-            "write: ino={}, fh={}, offset={}, len={}",
-            ino,
-            file_handle,
-            offset,
-            data.len()
-        );
-
-        if self.read_only {
-            return Err(libc::EROFS);
-        }
-
-        // Resolve write target: Local = staging file on disk (--advanced-writes),
-        // Streaming = append-only channel to CAS (default mode).
-        // Clone out of the map so we release the RwLock before doing I/O.
-        enum WriteTarget {
-            Local { file: Arc<Mutex<File>>, ino: u64 },
-            Streaming { ino: u64, channel: Arc<StreamingChannel> },
-        }
-
-        let target = {
-            let files = self.open_files.read().expect("open_files poisoned");
-            match files.get(&file_handle) {
-                Some(OpenFile::Local {
-                    ino,
-                    file,
-                    writable: true,
-                }) => WriteTarget::Local {
-                    file: Arc::clone(file),
-                    ino: *ino,
-                },
-                Some(OpenFile::Streaming { ino, channel }) => WriteTarget::Streaming {
-                    ino: *ino,
-                    channel: channel.clone(),
-                },
-                // Read-only handle, lazy handle, or unknown fh
-                _ => return Err(libc::EBADF),
-            }
-        };
-
-        match target {
-            WriteTarget::Local { file, ino: handle_ino } => {
-                let mut file = file.lock().expect("local file mutex poisoned");
-                file.seek(SeekFrom::Start(offset)).map_err(|_| libc::EIO)?;
-                file.write_all(data).map_err(|_| libc::EIO)?;
-                let written = data.len() as u32;
-                let new_end = offset + written as u64;
-                let mut inodes = self.inode_table.write().expect("inodes poisoned");
-                if let Some(entry) = inodes.get_mut(handle_ino) {
-                    if new_end > entry.size {
-                        if let Some(sd) = self.staging.dir() {
-                            sd.resize_bytes(entry.size, new_end);
-                        }
-                        entry.size = new_end;
-                    }
-                    entry.set_dirty();
-                }
-                inodes.touch(handle_ino);
-                Ok(written)
-            }
-            WriteTarget::Streaming {
-                ino: handle_ino,
-                channel,
-            } => {
-                // Check for previous worker error
-                if channel.error.lock().expect("error poisoned").is_some() {
-                    return Err(libc::EIO);
-                }
-
-                // Enforce append-only: offset must match bytes written so far
-                let expected = channel.bytes_written.load(Ordering::Relaxed);
-                if offset != expected {
-                    debug!(
-                        "streaming write: non-sequential offset={} (expected {}), ino={}",
-                        offset, expected, handle_ino
-                    );
-                    return Err(libc::EINVAL);
-                }
-
-                let len = data.len();
-                channel.tx.blocking_send(WriteMsg::Data(data.to_vec())).map_err(|_| {
-                    error!("streaming channel closed for ino={}", handle_ino);
-                    libc::EIO
-                })?;
-                channel.bytes_written.fetch_add(len as u64, Ordering::Relaxed);
-
-                let new_size = offset + len as u64;
-                let mut inodes = self.inode_table.write().expect("inodes poisoned");
-                if let Some(entry) = inodes.get_mut(handle_ino) {
-                    entry.size = new_size;
-                }
-
-                Ok(len as u32)
-            }
-        }
-    }
-
-    pub async fn flush(&self, ino: u64, file_handle: u64, pid: Option<u32>) -> VirtualFsResult<()> {
-        debug!("flush: ino={}, fh={}, pid={:?}", ino, file_handle, pid);
-
-        // Check if this is a streaming handle → synchronous upload + commit
-        let streaming_channel = {
-            let files = self.open_files.read().expect("open_files poisoned");
-            match files.get(&file_handle) {
-                Some(OpenFile::Streaming { channel, .. }) => Some(channel.clone()),
-                _ => None,
-            }
-        };
-
-        if let Some(channel) = streaming_channel {
-            // Check for worker errors first.
-            if channel.error.lock().expect("error poisoned").is_some() {
-                return Err(libc::EIO);
-            }
-
-            // Check current commit state.
-            {
-                let state = channel.state.lock().expect("state poisoned");
-                match &*state {
-                    CommitState::Committed => return Ok(()),
-                    CommitState::Failed(msg) => {
-                        debug!("flush: already failed for ino={}: {}", ino, msg);
-                        return Err(libc::EIO);
-                    }
-                    CommitState::Writing | CommitState::Deferred => {}
-                }
-            }
-
-            // PID-aware deferral: if the flushing process isn't the opener,
-            // this is a dup'd fd (e.g. shell redirection) — defer to release().
-            if let (Some(open_pid), Some(flush_pid)) = (channel.open_pid, pid)
-                && !same_process(open_pid, flush_pid)
-            {
-                debug!(
-                    "flush: deferring commit for ino={} (open_pid={}, flush_pid={})",
-                    ino, open_pid, flush_pid
-                );
-                self.install_commit_hook(ino, &channel);
-                *channel.state.lock().expect("state poisoned") = CommitState::Deferred;
-                return Ok(());
-            }
-
-            // Secondary gate: skip commit if no data was written (covers NFS
-            // and zero-write cases like `touch`). release() will handle it.
-            if channel.bytes_written.load(Ordering::Relaxed) == 0 {
-                self.install_commit_hook(ino, &channel);
-                *channel.state.lock().expect("state poisoned") = CommitState::Deferred;
-                return Ok(());
-            }
-
-            // Install hook before commit so concurrent open() can wait on us.
-            self.install_commit_hook(ino, &channel);
-
-            match self.streaming_commit(ino, &channel).await {
-                Ok(()) => {
-                    *channel.state.lock().expect("state poisoned") = CommitState::Committed;
-                    self.fulfill_commit_hook(ino, &channel, Ok(()));
-                }
-                Err(e) => {
-                    // CAS upload may have succeeded — file_info is preserved
-                    // in pending_info for retry in release(). Don't fulfill the
-                    // hook here: release() will retry and publish the final outcome.
-                    // Keeping the hook active ensures concurrent open(O_TRUNC) waits
-                    // for release() instead of racing with the retry.
-                    return Err(e);
-                }
-            }
-            return Ok(());
-        }
-
-        // Advanced writes mode: check if a previous async flush failed
-        if let Some(fm) = &self.flush_manager
-            && let Some(err_msg) = fm.check_error(ino)
-        {
-            error!("Deferred flush error for ino={}: {}", ino, err_msg);
-            return Err(libc::EIO);
-        }
-        Ok(())
-    }
-
-    pub async fn release(&self, file_handle: u64) -> VirtualFsResult<()> {
-        debug!("release: fh={}", file_handle);
-
-        let removed = self
-            .open_files
-            .write()
-            .expect("open_files poisoned")
-            .remove(&file_handle);
-
-        let released_ino = match &removed {
-            Some(OpenFile::Local { ino, .. })
-            | Some(OpenFile::Lazy { ino, .. })
-            | Some(OpenFile::Streaming { ino, .. }) => Some(*ino),
-            _ => None,
-        };
-        if let Some(ino) = released_ino {
-            self.inode_table.read().expect("inodes poisoned").drop_open_handles(ino);
-        }
-
-        let mut release_error: Option<i32> = None;
-
-        match removed {
-            Some(OpenFile::Local {
-                ino, writable: true, ..
-            }) => {
-                // Advanced writes: enqueue for async flush (skip unlinked files —
-                // user deleted the file, no point uploading it to remote).
-                let is_unlinked = self
-                    .inode_table
-                    .read()
-                    .expect("inodes poisoned")
-                    .get(ino)
-                    .is_some_and(|e| e.nlink == 0);
-                if !is_unlinked && let Some(fm) = &self.flush_manager {
-                    fm.enqueue(ino);
-                }
-            }
-            Some(OpenFile::Streaming { ino, channel }) => {
-                let needs_commit = {
-                    let state = channel.state.lock().expect("state poisoned");
-                    matches!(&*state, CommitState::Writing | CommitState::Deferred)
-                };
-                if needs_commit {
-                    // If flush() didn't defer (e.g. Writing state from a direct
-                    // release without flush), install the hook now.
-                    if channel.commit_hook.lock().expect("commit_hook poisoned").is_none() {
-                        self.install_commit_hook(ino, &channel);
-                    }
-
-                    let result = match self.streaming_commit(ino, &channel).await {
-                        Ok(()) => Ok(()),
-                        Err(e) => {
-                            // Retry only if CAS upload succeeded but Hub commit failed
-                            // (pending_info is preserved). If the worker itself died,
-                            // there's nothing to retry -- the data is gone.
-                            if channel.pending_info.lock().expect("pending_info poisoned").is_some() {
-                                tokio::time::sleep(Duration::from_secs(1)).await;
-                                self.streaming_commit(ino, &channel).await
-                            } else {
-                                Err(e)
-                            }
-                        }
-                    };
-
-                    match &result {
-                        Ok(()) => {
-                            *channel.state.lock().expect("state poisoned") = CommitState::Committed;
-                        }
-                        Err(e) => {
-                            let is_unlinked = self
-                                .inode_table
-                                .read()
-                                .expect("inodes poisoned")
-                                .get(ino)
-                                .is_none_or(|entry| entry.nlink == 0);
-                            if is_unlinked {
-                                debug!("streaming commit failed for unlinked ino={} (expected)", ino);
-                            } else {
-                                error!("DATA LOSS: streaming commit failed for ino={}: errno={}", ino, e);
-                            }
-                            self.revert_inode(ino, &channel.snapshot);
-                            *channel.state.lock().expect("state poisoned") =
-                                CommitState::Failed("commit failed".into());
-                            release_error = Some(*e);
-                        }
-                    }
-
-                    self.fulfill_commit_hook(ino, &channel, result);
-                }
-            }
-            _ => {}
-        }
-
-        if let Some(ino) = released_ino
-            && !self.has_open_handles(ino)
-        {
-            let (removed, is_clean, overlay_path) = {
-                let mut inodes = self.inode_table.write().expect("inodes poisoned");
-                // Capture path before removal so we can clean up the overlay
-                // backing file for an unlink-while-open case (POSIX delete on
-                // last close): unlink() saw open handles and skipped the
-                // overlay-side remove, so we have to do it here.
-                let overlay_path = self
-                    .overlay_backing
-                    .is_some()
-                    .then(|| inodes.get(ino).map(|e| e.full_path.clone()))
-                    .flatten();
-                let orphan = inodes.remove_orphan(ino);
-                let evicted = inodes.take_evict_pending(ino) && inodes.evict_if_safe(ino);
-                let removed = orphan || evicted;
-                let is_clean = !removed && inodes.get(ino).is_some_and(|entry| !entry.is_dirty());
-                (removed, is_clean, overlay_path.filter(|_| removed))
-            };
-            if removed {
-                if let Some(path) = overlay_path {
-                    if let Err(e) = self.remove_local_backing_file(ino, &path)
-                        && e.kind() != std::io::ErrorKind::NotFound
-                    {
-                        warn!("Failed to remove overlay file for ino={} on release: {}", ino, e);
-                    }
-                } else {
-                    self.drop_staging(ino);
-                }
-            } else if is_clean && self.staging.gc_one(ino, &self.inode_table).await {
-                debug!("staging GC: removed ino={} on release", ino);
-            }
-        }
-
-        // Per-inode staging locks are intentionally not cleaned up here:
-        // removing while another open() may hold the Arc would break
-        // serialization. Entries are tiny and bounded by inodes ever staged.
-
-        match release_error {
-            Some(e) => Err(e),
-            None => Ok(()),
-        }
-    }
-
-    /// Revert an inode to its pre-write state after a failed streaming commit.
-    fn revert_inode(&self, ino: u64, snapshot: &InodeSnapshot) {
-        let mut inodes = self.inode_table.write().expect("inodes poisoned");
-        if !snapshot.existed_before {
-            // File was created in this session — remove it entirely.
-            inodes.remove(ino);
-            info!("Reverted ino={}: removed (was newly created)", ino);
-        } else if let Some(entry) = inodes.get_mut(ino) {
-            // Overwrite — restore to pre-truncate state.
-            entry.xet_hash = snapshot.xet_hash.clone();
-            entry.size = snapshot.size;
-            entry.mtime = snapshot.mtime;
-            entry.pending_deletes = snapshot.pending_deletes.clone();
-            entry.dirty_generation = 0;
-            info!(
-                "Reverted ino={}: restored (hash={:?}, size={})",
-                ino, snapshot.xet_hash, snapshot.size
-            );
-        }
-    }
-
-    /// Finalize a streaming write: send Finish to the worker, await CAS upload, commit to Hub.
-    async fn streaming_commit(&self, ino: u64, channel: &StreamingChannel) -> Result<(), i32> {
-        assert!(!self.overlay(), "overlay forces advanced_writes; streaming unreachable");
-
-        // Unlinked files (nlink=0) must not be re-committed on close —
-        // user deleted the file, uploading would resurrect it.
-        if self
-            .inode_table
-            .read()
-            .expect("inodes poisoned")
-            .get(ino)
-            .is_some_and(|e| e.nlink == 0)
-        {
-            debug!("streaming_commit: skipping unlinked ino={}", ino);
-            return Ok(());
-        }
-
-        let file_info = {
-            let pending = channel.pending_info.lock().expect("pending_info poisoned").take();
-            if let Some(info) = pending {
-                info
-            } else {
-                let (result_tx, result_rx) = tokio::sync::oneshot::channel();
-                if channel.tx.send(WriteMsg::Finish(result_tx)).await.is_err() {
-                    // Channel closed: only treat as success if already committed.
-                    // If the worker died from an error, this is a real failure.
-                    let already_committed =
-                        matches!(&*channel.state.lock().expect("state poisoned"), CommitState::Committed);
-                    if already_committed {
-                        debug!("streaming_commit: channel closed but already committed for ino={}", ino);
-                        return Ok(());
-                    }
-                    error!("streaming_commit: channel closed with no prior commit for ino={}", ino);
-                    return Err(libc::EIO);
-                }
-                match result_rx.await {
-                    Ok(Ok(info)) => info,
-                    Ok(Err(e)) => {
-                        error!("Streaming upload failed for ino={}: {}", ino, e);
-                        return Err(libc::EIO);
-                    }
-                    Err(_) => {
-                        error!("Streaming worker dropped for ino={}", ino);
-                        return Err(libc::EIO);
-                    }
-                }
-            }
-        };
-
-        // Commit to Hub
-        let (full_path, pending_deletes) = {
-            let inodes = self.inode_table.read().expect("inodes poisoned");
-            let entry = inodes.get(ino).ok_or(libc::ENOENT)?;
-            (entry.full_path.to_string(), entry.pending_deletes.clone())
-        };
-
-        let mtime_ms = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_millis() as u64;
-
-        let mut ops: Vec<BatchOp> = vec![BatchOp::AddFile {
-            path: full_path.clone(),
-            xet_hash: file_info.hash().to_string(),
-            mtime: mtime_ms,
-            content_type: None,
-        }];
-        for old_path in &pending_deletes {
-            ops.push(BatchOp::DeleteFile { path: old_path.clone() });
-        }
-
-        if let Err(e) = self.hub_client.batch_operations(&ops).await {
-            error!("Failed to commit file {}: {}", full_path, e);
-            // CAS upload succeeded — preserve file_info for retry in release()
-            *channel.pending_info.lock().expect("pending_info poisoned") = Some(file_info);
-            return Err(libc::EIO);
-        }
-
-        let mut inodes = self.inode_table.write().expect("inodes poisoned");
-        if let Some(entry) = inodes.get_mut(ino) {
-            entry.apply_commit(
-                file_info.hash(),
-                file_info.file_size().expect("upload returned XetFileInfo without size"),
-                channel.dirty_generation_at_open.load(Ordering::Relaxed),
-            );
-        }
-
-        info!(
-            "Committed file: {} (hash={}, size={})",
-            full_path,
-            file_info.hash(),
-            file_info.file_size().expect("upload returned XetFileInfo without size"),
-        );
-
-        Ok(())
-    }
-
-    pub async fn create(
-        &self,
-        parent: u64,
-        name: &str,
-        mode: u16,
-        caller_uid: u32,
-        caller_gid: u32,
-        pid: Option<u32>,
-    ) -> VirtualFsResult<(VirtualFsAttr, u64)> {
-        if self.read_only {
-            return Err(libc::EROFS);
-        }
-        if self.filter_os_files && is_os_junk(name) {
-            debug!("create: rejecting OS junk file: {}", name);
-            return Err(libc::EACCES);
-        }
-
-        debug!("create: parent={}, name={}", parent, name);
-
-        // Fetch remote children so we can detect name collisions (also validates parent)
-        self.ensure_children_loaded(parent).await?;
-
-        let now = SystemTime::now();
-        // Re-check parent + EEXIST + insert under one write lock (TOCTOU guard)
-        let (ino, full_path, dirty_gen) = {
-            let mut inodes = self.inode_table.write().expect("inodes poisoned");
-            let parent_entry = match inodes.get(parent) {
-                Some(e) if e.kind == InodeKind::Directory => e,
-                Some(_) => return Err(libc::ENOTDIR),
-                None => return Err(libc::ENOENT),
-            };
-            let full_path = inode::child_path(&parent_entry.full_path, name);
-            if inodes.lookup_child(parent, name).is_some() {
-                return Err(libc::EEXIST);
-            }
-            // Insert as dirty — content lives in local staging until flushed
-            let ino = inodes.insert(
-                parent,
-                name.to_string(),
-                full_path.clone(),
-                InodeKind::File,
-                0,
-                now,
-                None,
-                mode,
-                caller_uid,
-                caller_gid,
-            );
-            let dirty_gen = if let Some(entry) = inodes.get_mut(ino) {
-                entry.set_dirty();
-                entry.dirty_generation
-            } else {
-                0
-            };
-            inodes.touch_parent(parent, now);
-            inodes.touch(ino);
-            (ino, full_path, dirty_gen)
-        };
-
-        self.negative_cache_remove(&full_path);
-        if let Some(fm) = &self.flush_manager {
-            fm.cancel_delete(&full_path);
-        }
-
-        if self.advanced_writes {
-            // Advanced mode: staging file on disk + async flush (or overlay file in overlay mode).
-            match self.open_local_backing_file(ino, &full_path, true, true, true, true) {
-                Ok(file) => {
-                    if let Err(e) = self.set_local_backing_mode(&full_path, mode) {
-                        error!("Failed to set local backing mode for {}: {}", full_path, e);
-                        self.inode_table.write().expect("inodes poisoned").remove(ino);
-                        return Err(libc::EIO);
-                    }
-                    let file_handle = self.alloc_file_handle();
-                    let inodes = self.inode_table.read().expect("inodes poisoned");
-                    inodes.bump_open_handles(ino);
-                    self.open_files.write().expect("open_files poisoned").insert(
-                        file_handle,
-                        OpenFile::Local {
-                            ino,
-                            file: Arc::new(Mutex::new(file)),
-                            writable: true,
-                        },
-                    );
-
-                    let attr = self.make_vfs_attr(inodes.get(ino).ok_or(libc::ENOENT)?);
-                    Ok((attr, file_handle))
-                }
-                Err(e) => {
-                    error!("Failed to create staging file: {}", e);
-                    self.inode_table.write().expect("inodes poisoned").remove(ino);
-                    Err(libc::EIO)
-                }
-            }
-        } else {
-            // Simple mode: streaming writer with channel-based decoupling.
-            let snapshot = InodeSnapshot {
-                xet_hash: None,
-                size: 0,
-                mtime: now,
-                pending_deletes: Vec::new(),
-                existed_before: false,
-            };
-            let (file_handle, channel) = match self.setup_streaming_writer(pid, snapshot, dirty_gen).await {
-                Ok(r) => r,
-                Err(e) => {
-                    self.inode_table.write().expect("inodes poisoned").remove(ino);
-                    return Err(e);
-                }
-            };
-
-            let inodes = self.inode_table.read().expect("inodes poisoned");
-            inodes.bump_open_handles(ino);
-            self.open_files
-                .write()
-                .expect("open_files poisoned")
-                .insert(file_handle, OpenFile::Streaming { ino, channel });
-
-            let attr = self.make_vfs_attr(inodes.get(ino).ok_or(libc::ENOENT)?);
-            Ok((attr, file_handle))
-        }
-    }
-
-    pub async fn mkdir(
-        &self,
-        parent: u64,
-        name: &str,
-        mode: u16,
-        caller_uid: u32,
-        caller_gid: u32,
-    ) -> VirtualFsResult<VirtualFsAttr> {
-        if self.read_only {
-            return Err(libc::EROFS);
-        }
-        if self.filter_os_files && is_os_junk(name) {
-            debug!("mkdir: rejecting OS junk directory: {}", name);
-            return Err(libc::EACCES);
-        }
-
-        debug!("mkdir: parent={}, name={}", parent, name);
-
-        // Fetch remote children so we can detect name collisions (also validates parent)
-        self.ensure_children_loaded(parent).await?;
-
-        let now = SystemTime::now();
-        // Re-check parent + EEXIST + insert under one write lock (TOCTOU guard)
-        let (ino, full_path) = {
-            let mut inodes = self.inode_table.write().expect("inodes poisoned");
-            let parent_entry = match inodes.get(parent) {
-                Some(e) if e.kind == InodeKind::Directory => e,
-                Some(_) => return Err(libc::ENOTDIR),
-                None => return Err(libc::ENOENT),
-            };
-            let full_path = inode::child_path(&parent_entry.full_path, name);
-            if inodes.lookup_child(parent, name).is_some() {
-                return Err(libc::EEXIST);
-            }
-            // New dir starts with children_loaded=true (empty, nothing to fetch)
-            let ino = inodes.insert(
-                parent,
-                name.to_string(),
-                full_path.clone(),
-                InodeKind::Directory,
-                0,
-                now,
-                None,
-                mode,
-                caller_uid,
-                caller_gid,
-            );
-            if let Some(entry) = inodes.get_mut(ino) {
-                entry.children_loaded_at = Some(Instant::now());
-                // children_from_remote stays false: a freshly-mkdir'd dir has
-                // no remote presence yet, so lookup-miss can serve ENOENT
-                // without HEAD/list_tree probes.
-                if self.overlay() {
-                    entry.set_dirty();
-                }
-            }
-            // nlink already incremented by insert()
-            inodes.touch_parent(parent, now);
-            (ino, full_path)
-        };
-
-        self.negative_cache_remove(&full_path);
-
-        // Overlay: persist directory to disk.
-        if let Some(overlay) = &self.overlay_backing {
-            if let Err(e) = overlay.create_parent_dirs(&full_path) {
-                error!("Failed to create overlay parent directories for {}: {}", full_path, e);
-                self.inode_table.write().expect("inodes poisoned").remove(ino);
-                return Err(libc::EIO);
-            }
-            if let Err(e) = overlay.create_dir(&full_path, mode) {
-                error!("Failed to create overlay directory {}: {}", full_path, e);
-                self.inode_table.write().expect("inodes poisoned").remove(ino);
-                return Err(e.raw_os_error().unwrap_or(libc::EIO));
-            }
-        }
-
-        let inodes = self.inode_table.read().expect("inodes poisoned");
-        Ok(self.make_vfs_attr(inodes.get(ino).ok_or(libc::ENOENT)?))
-    }
-
-    pub async fn unlink(&self, parent: u64, name: &str) -> VirtualFsResult<()> {
-        if self.read_only {
-            return Err(libc::EROFS);
-        }
-
-        debug!("unlink: parent={}, name={}", parent, name);
-
-        self.ensure_children_loaded(parent).await?;
-
-        let (ino, full_path, needs_remote_delete) = {
-            let inodes = self.inode_table.read().expect("inodes poisoned");
-            let entry = match inodes.lookup_child(parent, name) {
-                Some(entry) if entry.kind != InodeKind::Directory => entry,
-                Some(_) => return Err(libc::EISDIR),
-                None => return Err(libc::ENOENT),
-            };
-            // Overlay: cannot delete clean remote entries (no whiteout support).
-            // Deleting dirty (local) entries is allowed; remote reappears on remount.
-            if self.is_overlay_immutable(entry) {
-                return Err(libc::EPERM);
-            }
-            // Remote delete only when last link is removed and file exists on the hub.
-            // Skipped in overlay mode (writes never propagate to remote).
-            let needs_remote = !self.overlay() && entry.xet_hash.is_some() && entry.nlink <= 1;
-            (entry.inode, entry.full_path.to_string(), needs_remote)
-        };
-
-        // In streaming mode, block unlink while the file has any open handles.
-        // This prevents editors (vim) from deleting a file they can't rewrite
-        // (streaming mode returns EPERM for O_RDWR without O_TRUNC), which
-        // would cause silent data loss. Same approach as mountpoint-s3.
-        if !self.advanced_writes && self.has_open_handles(ino) {
-            debug!("unlink: blocked for ino={} (has open handles in streaming mode)", ino);
-            return Err(libc::EPERM);
-        }
-
-        // Advanced writes: queue delete for batched flush in the flush_loop.
-        // Simple mode: delete synchronously (one HTTP call per unlink).
-        if needs_remote_delete {
-            if let Some(fm) = &self.flush_manager {
-                fm.enqueue_delete(full_path.clone());
-            } else if let Err(e) = self
-                .hub_client
-                .batch_operations(&[BatchOp::DeleteFile {
-                    path: full_path.clone(),
-                }])
-                .await
-            {
-                error!("Remote delete failed for {}: {}", full_path, e);
-                return Err(libc::EIO);
-            }
-        }
-
-        // Remote succeeded (or no remote needed) — now unlink locally.
-        // unlink_one decrements nlink; the inode stays in the table with nlink=0
-        // so open file handles can still fstat() it.
-        let inode_fully_removed = {
-            let mut inodes = self.inode_table.write().expect("inodes poisoned");
-            let last_link = inodes
-                .unlink_one(parent, name)
-                .map(|(removed, _)| removed)
-                .unwrap_or(false);
-            // Update parent mtime/ctime (POSIX: directory was modified)
-            let now = SystemTime::now();
-            inodes.touch_parent(parent, now);
-            // If last link gone and no open handles, remove the orphan immediately.
-            // Otherwise leave it for release() so an open fd can keep writing
-            // (POSIX unlink-while-open) without us yanking the staging file +
-            // debiting the budget twice.
-            let no_handles = !inodes.has_open_handles(ino);
-            if last_link && no_handles {
-                inodes.remove_orphan(ino);
-            }
-            last_link && no_handles
-        };
-
-        // Seed the negative cache before any await: with the inode already
-        // gone from the table, a concurrent lookup under a stale parent would
-        // otherwise HEAD the still-existing remote (the delete is only queued)
-        // and resurrect this name locally during the drop_locked await window.
-        self.negative_cache_insert(full_path.clone());
-
-        // Clean up staging file only when the inode is actually gone. With an
-        // open fd, drop_staging waits until release() so writes through the
-        // surviving fd can still update bytes_used coherently.
-        if inode_fully_removed {
-            if self.overlay() {
-                if let Err(e) = self.remove_local_backing_file(ino, &full_path)
-                    && e.kind() != std::io::ErrorKind::NotFound
-                {
-                    warn!("Failed to remove overlay file for ino={}: {}", ino, e);
-                }
-            } else {
-                self.staging.drop_locked(ino).await;
-            }
-        }
-
-        info!("Deleted file: {}", full_path);
-        Ok(())
-    }
-
-    pub async fn symlink(
-        &self,
-        parent: u64,
-        name: &str,
-        target: &str,
-        mode: u16,
-        caller_uid: u32,
-        caller_gid: u32,
-    ) -> VirtualFsResult<VirtualFsAttr> {
-        if self.read_only {
-            return Err(libc::EROFS);
-        }
-
-        debug!("symlink: parent={}, name={}, target={}", parent, name, target);
-
-        self.ensure_children_loaded(parent).await?;
-
-        let now = SystemTime::now();
-        let (ino, full_path) = {
-            let mut inodes = self.inode_table.write().expect("inodes poisoned");
-            let parent_entry = match inodes.get(parent) {
-                Some(e) if e.kind == InodeKind::Directory => e,
-                Some(_) => return Err(libc::ENOTDIR),
-                None => return Err(libc::ENOENT),
-            };
-            let full_path = inode::child_path(&parent_entry.full_path, name);
-            if inodes.lookup_child(parent, name).is_some() {
-                return Err(libc::EEXIST);
-            }
-            let ino = inodes.insert(
-                parent,
-                name.to_string(),
-                full_path.clone(),
-                InodeKind::Symlink,
-                target.len() as u64,
-                now,
-                None,
-                mode,
-                caller_uid,
-                caller_gid,
-            );
-            if let Some(entry) = inodes.get_mut(ino) {
-                entry.symlink_target = Some(target.to_string());
-            }
-            inodes.touch_parent(parent, now);
-            (ino, full_path)
-        };
-
-        self.negative_cache_remove(&full_path);
-        let inodes = self.inode_table.read().expect("inodes poisoned");
-        Ok(self.make_vfs_attr(inodes.get(ino).ok_or(libc::ENOENT)?))
-    }
-
-    pub fn readlink(&self, ino: u64) -> VirtualFsResult<String> {
-        let inodes = self.inode_table.read().expect("inodes poisoned");
-        match inodes.get(ino) {
-            Some(entry) if entry.kind == InodeKind::Symlink => entry.symlink_target.clone().ok_or(libc::EINVAL),
-            Some(_) => Err(libc::EINVAL),
-            None => Err(libc::ENOENT),
-        }
-    }
-
-    pub async fn link(&self, _ino: u64, _new_parent: u64, _new_name: &str) -> VirtualFsResult<VirtualFsAttr> {
-        // Hard links are not supported — they are ephemeral (in-memory only) and never
-        // persisted to the hub, which makes them a source of subtle bugs with no benefit.
-        Err(libc::ENOTSUP)
-    }
-
-    pub async fn rmdir(&self, parent: u64, name: &str) -> VirtualFsResult<()> {
-        if self.read_only {
-            return Err(libc::EROFS);
-        }
-
-        debug!("rmdir: parent={}, name={}", parent, name);
-
-        // Flush any queued remote deletes (e.g. from rm -rf that unlinked
-        // files before calling rmdir on the now-empty directory).
-        if let Some(fm) = &self.flush_manager {
-            fm.flush_deletes().await;
-        }
-
-        self.ensure_children_loaded(parent).await?;
-
-        let (ino, full_path) = {
-            let inodes = self.inode_table.read().expect("inodes poisoned");
-            match inodes.lookup_child(parent, name) {
-                Some(entry) if entry.kind == InodeKind::Directory => {
-                    // Overlay: cannot remove clean remote directories.
-                    if self.is_overlay_immutable(entry) {
-                        return Err(libc::EPERM);
-                    }
-                    if !entry.children.is_empty() {
-                        return Err(libc::ENOTEMPTY);
-                    }
-                    (entry.inode, entry.full_path.to_string())
-                }
-                Some(_) => return Err(libc::ENOTDIR),
-                None => return Err(libc::ENOENT),
-            }
-        };
-
-        // Load remote children so we don't miss any before the emptiness check
-        self.ensure_children_loaded(ino).await?;
-
-        // Overlay: remove on-disk dir before mutating the inode table so a
-        // failure can't leave the inode tree out of sync with the backing.
-        if let Some(overlay) = &self.overlay_backing
-            && let Err(e) = overlay.remove_dir(&full_path)
-            && e.kind() != std::io::ErrorKind::NotFound
-        {
-            return Err(e.raw_os_error().unwrap_or(libc::EIO));
-        }
-
-        // Re-check ENOTEMPTY + remove under a single write lock to prevent
-        // a concurrent create/mkdir from inserting a child in between.
-        {
-            let mut inodes = self.inode_table.write().expect("inodes poisoned");
-            match inodes.get(ino) {
-                Some(entry) if !entry.children.is_empty() => return Err(libc::ENOTEMPTY),
-                None => return Err(libc::ENOENT),
-                _ => {}
-            }
-            inodes.remove(ino);
-            // nlink adjusted by remove()
-            inodes.touch_parent(parent, SystemTime::now());
-        }
-        self.negative_cache_insert(full_path);
-        Ok(())
-    }
-
-    pub async fn rename(
-        &self,
-        parent: u64,
-        name: &str,
-        newparent: u64,
-        newname: &str,
-        no_replace: bool,
-    ) -> VirtualFsResult<()> {
-        if self.read_only {
-            return Err(libc::EROFS);
-        }
-        if self.filter_os_files && is_os_junk(newname) {
-            debug!("rename: rejecting rename to OS junk name: {}", newname);
-            return Err(libc::EACCES);
-        }
-
-        debug!(
-            "rename: parent={}, name={}, newparent={}, newname={}",
-            parent, name, newparent, newname
-        );
-
-        // Noop: renaming to same location
-        if parent == newparent && name == newname {
-            return Ok(());
-        }
-
-        self.ensure_children_loaded(parent).await?;
-        if parent != newparent {
-            self.ensure_children_loaded(newparent).await?;
-        }
-
-        // Overlay: cannot rename clean remote entries (no whiteout for old path).
-        if let Some(src) = self
-            .inode_table
-            .read()
-            .expect("inodes poisoned")
-            .lookup_child(parent, name)
-            && self.is_overlay_immutable(src)
-        {
-            return Err(libc::EPERM);
-        }
-
-        // Pre-load lazy directories before the sync validate phase:
-        // - Source subtree: so descendant_files is complete for remote rename ops
-        // - Destination dir (if exists): so children.is_empty() check is accurate
-        let (src_dir, dst_dir) = {
-            let inodes = self.inode_table.read().expect("inodes poisoned");
-            let src = inodes
-                .lookup_child(parent, name)
-                .filter(|e| e.kind == InodeKind::Directory)
-                .map(|e| e.inode);
-            let dst = inodes
-                .lookup_child(newparent, newname)
-                .filter(|e| e.kind == InodeKind::Directory)
-                .map(|e| e.inode);
-            (src, dst)
-        };
-        if let Some(src_ino) = src_dir {
-            self.ensure_subtree_loaded(src_ino).await?;
-            // Overlay shadow guard: a local dir that shares its name with a
-            // remote dir is dirty (set by merge_overlay_entries), so the
-            // top-level immutable check above lets it through. After loading
-            // the subtree we can spot remote descendants — moving them
-            // locally without a remote rename would leave them dangling on
-            // the new local path while the remote tree keeps the old one.
-            if self.overlay()
-                && self
-                    .inode_table
-                    .read()
-                    .expect("inodes poisoned")
-                    .has_clean_descendants(src_ino)
-            {
-                return Err(libc::EPERM);
-            }
-        }
-        if let Some(dst_ino) = dst_dir {
-            self.ensure_children_loaded(dst_ino).await?;
-        }
-
-        // Phase 1: validate under read lock, collect everything we need
-        let info = self.rename_validate(parent, name, newparent, newname, no_replace)?;
-
-        // POSIX: rename(a, b) where a and b are hard links to the same inode is a no-op.
-        // Check before remote sync to avoid spurious backend mutations.
-        {
-            let inodes = self.inode_table.read().expect("inodes poisoned");
-            if inodes
-                .lookup_child(newparent, newname)
-                .is_some_and(|e| e.inode == info.ino)
-            {
-                return Ok(());
-            }
-        }
-
-        // Phase 2: sync to remote (add + delete ops). Skipped in overlay mode
-        // since writes never propagate beyond the local backing.
-        let remote_mutated = if self.overlay() {
-            false
-        } else {
-            self.rename_remote(&info).await?
-        };
-        let old_path = info.old_path.clone();
-
-        // Overlay: move the on-disk file to match the new path.
-        if let Some(overlay) = &self.overlay_backing {
-            let old_exists = overlay.exists(&info.old_path).map_err(|e| {
-                error!("Overlay rename: failed to stat {}: {}", info.old_path, e);
-                libc::EIO
-            })?;
-            if !old_exists {
-                return Err(libc::EPERM);
-            }
-            if let Err(e) = overlay.create_parent_dirs(&info.new_full_path) {
-                error!("Overlay rename: failed to create destination parents: {}", e);
-                return Err(e.raw_os_error().unwrap_or(libc::EIO));
-            }
-            if let Err(e) = overlay.rename(&info.old_path, &info.new_full_path) {
-                error!(
-                    "Overlay rename {} -> {} failed: {}",
-                    info.old_path, info.new_full_path, e
-                );
-                return Err(e.raw_os_error().unwrap_or(libc::EIO));
-            }
-        }
-
-        // Phase 3: apply to local inode table under write lock.
-        // If Phase 2 mutated the remote and Phase 3 fails with ENOENT (source
-        // concurrently unlinked), swallow the error and let poll reconcile.
-        // Destination-conflict errors (EEXIST, EISDIR, etc.) are propagated
-        // since poll may not fix a dirty local inode at the destination path.
-        match self.rename_apply_local(info, parent, name, newparent, newname, no_replace) {
-            Ok(replaced_staging_ino) => {
-                // Seed the negative cache before any await: with the source
-                // already moved locally, a concurrent lookup of `old_path`
-                // would otherwise HEAD the still-existing remote object
-                // during the drop_locked await window and resurrect it.
-                self.negative_cache_insert(old_path);
-                if let Some(ino) = replaced_staging_ino {
-                    self.staging.drop_locked(ino).await;
-                }
-                Ok(())
-            }
-            Err(libc::ENOENT) if remote_mutated || self.overlay() => {
-                warn!("rename: source gone after rename; next dir load will reconcile");
-                // Invalidate parents so the next lookup re-fetches from remote
-                // instead of trusting the stale children_loaded state.
-                let mut inodes = self.inode_table.write().expect("inodes poisoned");
-                inodes.invalidate_children(parent);
-                if parent != newparent {
-                    inodes.invalidate_children(newparent);
-                }
-                Ok(())
-            }
-            Err(errno) => Err(errno),
-        }
-    }
-
-    /// Phase 1: validate rename under inode read lock, return all info needed for phases 2+3.
-    fn rename_validate(
-        &self,
-        parent: u64,
-        name: &str,
-        newparent: u64,
-        newname: &str,
-        no_replace: bool,
-    ) -> VirtualFsResult<RenameInfo> {
-        let inodes = self.inode_table.read().expect("inodes poisoned");
-
-        let src = inodes.lookup_child(parent, name).ok_or(libc::ENOENT)?;
-
-        // Destination parent must exist
-        let new_parent_entry = inodes.get(newparent).ok_or(libc::ENOENT)?;
-        let new_full_path = inode::child_path(&new_parent_entry.full_path, newname);
-
-        // Prevent moving a directory into its own subtree (would create a cycle)
-        if src.kind == InodeKind::Directory {
-            let mut ancestor = newparent;
-            while ancestor != 1 {
-                if ancestor == src.inode {
-                    return Err(libc::EINVAL);
-                }
-                ancestor = match inodes.get(ancestor) {
-                    Some(e) => e.parent,
-                    None => break,
-                };
-            }
-        }
-
-        // Check destination conflicts
-        if let Some(existing) = inodes.lookup_child(newparent, newname) {
-            if no_replace {
-                return Err(libc::EEXIST);
-            }
-            match (src.kind, existing.kind) {
-                (InodeKind::File | InodeKind::Symlink, InodeKind::Directory) => return Err(libc::EISDIR),
-                (InodeKind::Directory, InodeKind::File | InodeKind::Symlink) => return Err(libc::ENOTDIR),
-                (InodeKind::Directory, InodeKind::Directory) if !existing.children.is_empty() => {
-                    return Err(libc::ENOTEMPTY);
-                }
-                _ => {}
-            }
-        }
-
-        // For directories, collect all descendant clean files for remote rename
-        let descendant_files = if src.kind == InodeKind::Directory {
-            let mut files = Vec::new();
-            let mut stack = vec![src.inode];
-            debug!(
-                "rename_validate: dir ino={} children_loaded={} children_count={}",
-                src.inode,
-                inodes.is_children_loaded(src.inode),
-                src.children.len(),
-            );
-            while let Some(dir_ino) = stack.pop() {
-                if let Some(entry) = inodes.get(dir_ino) {
-                    for child_ref in &entry.children {
-                        if let Some(child) = inodes.get(child_ref.ino) {
-                            debug!(
-                                "rename_validate: child ino={} path={} dirty={} xet_hash={:?}",
-                                child_ref.ino,
-                                child.full_path,
-                                child.is_dirty(),
-                                child.xet_hash.as_deref()
-                            );
-                            match child.kind {
-                                InodeKind::File if !child.is_dirty() && child.xet_hash.is_some() => {
-                                    files.push((
-                                        child.full_path.to_string(),
-                                        child.xet_hash.clone().expect("checked is_some above"),
-                                    ));
-                                }
-                                InodeKind::Directory => stack.push(child_ref.ino),
-                                _ => {}
-                            }
-                        }
-                    }
-                }
-            }
-            files
-        } else {
-            Vec::new()
-        };
-
-        Ok(RenameInfo {
-            ino: src.inode,
-            old_path: src.full_path.to_string(),
-            kind: src.kind,
-            xet_hash: src.xet_hash.clone(),
-            is_dirty: src.is_dirty(),
-            new_full_path,
-            descendant_files,
-        })
-    }
-
-    /// Phase 2: send batch rename ops to the Hub (add new paths + delete old ones).
-    /// Returns true if remote operations were actually sent, false if skipped
-    /// (e.g. dirty file with no remote presence).
-    async fn rename_remote(&self, info: &RenameInfo) -> VirtualFsResult<bool> {
-        let mtime_ms = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_millis() as u64;
-
-        let ops: Vec<BatchOp> = if info.kind == InodeKind::File
-            && !info.is_dirty
-            && let Some(ref hash) = info.xet_hash
-        {
-            vec![
-                BatchOp::AddFile {
-                    path: info.new_full_path.clone(),
-                    xet_hash: hash.clone(),
-                    mtime: mtime_ms,
-                    content_type: None,
-                },
-                BatchOp::DeleteFile {
-                    path: info.old_path.clone(),
-                },
-            ]
-        } else if info.kind == InodeKind::Directory && !info.descendant_files.is_empty() {
-            // Hub batch API requires all adds before all deletes
-            let mut ops = Vec::with_capacity(info.descendant_files.len() * 2);
-            let mut deletes = Vec::with_capacity(info.descendant_files.len());
-            for (child_old_path, child_hash) in &info.descendant_files {
-                let suffix = child_old_path.strip_prefix(&info.old_path).unwrap_or(child_old_path);
-                let child_new_path = format!("{}{}", info.new_full_path, suffix);
-                ops.push(BatchOp::AddFile {
-                    path: child_new_path,
-                    xet_hash: child_hash.clone(),
-                    mtime: mtime_ms,
-                    content_type: None,
-                });
-                deletes.push(BatchOp::DeleteFile {
-                    path: child_old_path.clone(),
-                });
-            }
-            ops.append(&mut deletes);
-            ops
-        } else {
-            return Ok(false);
-        };
-
-        debug!(
-            "rename_remote: {} -> {} ops={}",
-            info.old_path,
-            info.new_full_path,
-            ops.len()
-        );
-        if let Err(e) = self.hub_client.batch_operations(&ops).await {
-            error!("Failed to rename {} -> {}: {}", info.old_path, info.new_full_path, e);
-            return Err(libc::EIO);
-        }
-        debug!("rename_remote: success");
-        Ok(true)
-    }
-
-    /// Phase 3: apply rename to local inode table under write lock.
-    /// Returns `Ok(Some(ino))` when a staging file for a replaced target needs
-    /// to be dropped; the caller must do that asynchronously under the
-    /// per-inode staging lock to serialize with in-flight flush uploads.
-    fn rename_apply_local(
-        &self,
-        info: RenameInfo,
-        parent: u64,
-        oldname: &str,
-        newparent: u64,
-        newname: &str,
-        no_replace: bool,
-    ) -> VirtualFsResult<Option<u64>> {
-        self.negative_cache_remove(&info.new_full_path);
-        // Cancel any queued remote delete for the destination path (e.g. rm a && mv b a).
-        // For directories, also cancel descendant deletes (e.g. rm -rf dir && mv newdir dir).
-        if let Some(fm) = &self.flush_manager {
-            fm.cancel_delete(&info.new_full_path);
-            if info.kind == InodeKind::Directory {
-                fm.cancel_delete_prefix(&format!("{}/", info.new_full_path));
-            }
-        }
-
-        let mut inodes = self.inode_table.write().expect("inodes poisoned");
-
-        // Re-validate source still exists (could have been unlinked concurrently)
-        if inodes.get(info.ino).is_none() {
-            return Err(libc::ENOENT);
-        }
-
-        // Re-check destination under write lock (a concurrent create could have
-        // inserted one between phase 1 read-lock and this write-lock).
-        let replace_target = if let Some(existing) = inodes.lookup_child(newparent, newname) {
-            // POSIX: rename(a, b) where a and b are hard links to the same inode is a no-op
-            if existing.inode == info.ino {
-                return Ok(None);
-            }
-            if no_replace {
-                return Err(libc::EEXIST);
-            }
-            match (info.kind, existing.kind) {
-                // POSIX: non-directory cannot replace directory
-                (InodeKind::File | InodeKind::Symlink, InodeKind::Directory) => return Err(libc::EISDIR),
-                // POSIX: directory cannot replace non-directory
-                (InodeKind::Directory, InodeKind::File | InodeKind::Symlink) => return Err(libc::ENOTDIR),
-                (InodeKind::Directory, InodeKind::Directory) if !existing.children.is_empty() => {
-                    return Err(libc::ENOTEMPTY);
-                }
-                _ => {}
-            }
-            Some((existing.inode, existing.kind))
-        } else {
-            None
-        };
-        let mut replaced_staging_ino: Option<u64> = None;
-        if let Some((existing_ino, existing_kind)) = replace_target {
-            if existing_kind == InodeKind::Directory {
-                // Directories can't be hard-linked, so remove() is correct
-                // (remove() adjusts parent nlink for directories)
-                inodes.remove(existing_ino);
-            } else {
-                inodes.unlink_one(newparent, newname);
-                if !inodes.has_open_handles(existing_ino) && inodes.remove_orphan(existing_ino) {
-                    replaced_staging_ino = Some(existing_ino);
-                }
-            }
-        }
-
-        // Dirty file with a remote presence: record old path for deletion at flush time.
-        if info.is_dirty
-            && info.kind == InodeKind::File
-            && info.xet_hash.is_some()
-            && let Some(entry) = inodes.get_mut(info.ino)
-        {
-            entry.pending_deletes.push(info.old_path.clone());
-        }
-
-        // Dirty descendants of a renamed directory: record their old remote paths
-        // for deletion at flush time (clean descendants are handled in rename_remote).
-        if info.kind == InodeKind::Directory {
-            let mut stack = vec![info.ino];
-            while let Some(dir_ino) = stack.pop() {
-                let children: Vec<inode::DirChild> =
-                    inodes.get(dir_ino).map(|e| e.children.clone()).unwrap_or_default();
-                for child_ref in children {
-                    if let Some(child) = inodes.get(child_ref.ino) {
-                        match child.kind {
-                            InodeKind::File if child.is_dirty() && child.xet_hash.is_some() => {
-                                let old_path = child.full_path.to_string();
-                                if let Some(child_mut) = inodes.get_mut(child_ref.ino) {
-                                    child_mut.pending_deletes.push(old_path);
-                                }
-                            }
-                            InodeKind::Directory => stack.push(child_ref.ino),
-                            _ => {}
-                        }
-                    }
-                }
-            }
-        }
-
-        // Remove old path mapping for the renamed entry
-        inodes.remove_path(&info.old_path);
-
-        inodes.move_child(info.ino, parent, oldname, newparent, newname);
-        inodes.update_subtree_paths(info.ino, info.new_full_path);
-
-        // Update parent mtime/ctime for both old and new parents (POSIX: directories modified)
-        let now = SystemTime::now();
-        inodes.touch_parent(parent, now);
-        if parent != newparent {
-            inodes.touch_parent(newparent, now);
-        }
-        // Update source ctime (POSIX: inode metadata changed)
-        if let Some(e) = inodes.get_mut(info.ino) {
-            e.ctime = now;
-        }
-        drop(inodes);
-
-        Ok(replaced_staging_ino)
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    pub async fn setattr(
-        &self,
-        ino: u64,
-        size: Option<u64>,
-        mode: Option<u16>,
-        uid: Option<u32>,
-        gid: Option<u32>,
-        atime: Option<SystemTime>,
-        mtime: Option<SystemTime>,
-    ) -> VirtualFsResult<VirtualFsAttr> {
-        debug!("setattr: ino={}, size={:?}", ino, size);
-
-        // EROFS check blocks all changes on RO mounts
-        if self.read_only {
-            return Err(libc::EROFS);
-        }
-
-        if let Some(new_size) = size {
-            // Validate inode exists and is a file before any side effects
-            let full_path = {
-                let inodes = self.inode_table.read().expect("inodes poisoned");
-                match inodes.get(ino) {
-                    Some(e) if e.kind != InodeKind::File => return Err(libc::EISDIR),
-                    Some(e) => e.full_path.clone(),
-                    None => return Err(libc::ENOENT),
-                }
-            };
-
-            if !self.advanced_writes {
-                // Simple mode: ftruncate via setattr is silently ignored.
-                // Real truncation goes through open(O_TRUNC) which is handled separately.
-            } else {
-                // Advanced mode: truncation is applied to the staging file on disk
-                let staging_mutex = self.staging.lock(ino);
-                let _staging_guard = staging_mutex.lock().await;
-
-                let local_exists = self.local_backing_exists(ino, &full_path).map_err(|e| {
-                    error!("Failed to check local backing file for ino={}: {}", ino, e);
-                    e.raw_os_error().unwrap_or(libc::EIO)
-                })?;
-
-                if self.overlay() && !local_exists {
-                    return Err(libc::EPERM);
-                }
-
-                // GC accounting (non-overlay only): snapshot staging bytes before
-                // any mutation so the size delta is applied correctly at the end.
-                let old_staging_size = self
-                    .staging
-                    .dir()
-                    .filter(|_| !self.overlay())
-                    .map(|sd| sd.file_size(ino))
-                    .unwrap_or(0);
-
-                if !local_exists {
-                    if new_size > 0 {
-                        let staging_path = self
-                            .staging
-                            .path(ino)
-                            .expect("staging directory required for advanced writes");
-                        let (xet_hash, file_size) = {
-                            let inodes = self.inode_table.read().expect("inodes poisoned");
-                            let entry = inodes.get(ino).ok_or(libc::ENOENT)?;
-                            (entry.xet_hash.clone().unwrap_or_default(), entry.size)
-                        };
-                        if !xet_hash.is_empty() && file_size > 0 {
-                            if let Err(e) = self
-                                .xet_sessions
-                                .download_to_file(&xet_hash, file_size, &staging_path)
-                                .await
-                            {
-                                error!("Failed to download file for truncate: {}", e);
-                                return Err(libc::EIO);
-                            }
-                        } else if let Err(e) = self.open_local_backing_file(ino, &full_path, true, true, true, true) {
-                            error!("Failed to create staging file for truncate: {}", e);
-                            return Err(libc::EIO);
-                        }
-                    } else if let Err(e) = self.open_local_backing_file(ino, &full_path, true, true, true, true) {
-                        error!("Failed to create local backing file for truncate: {}", e);
-                        return Err(libc::EIO);
-                    }
-                }
-
-                // Apply the size change under the same write lock so write() cannot
-                // race between the local truncate and inode metadata update.
-                let mut inodes = self.inode_table.write().expect("inodes poisoned");
-                let size_result = if new_size == 0 {
-                    self.open_local_backing_file(ino, &full_path, true, true, true, true)
-                        .map(|_| ())
-                } else {
-                    self.open_local_backing_file(ino, &full_path, false, true, false, false)
-                        .and_then(|file| file.set_len(new_size))
-                };
-                if let Err(e) = size_result {
-                    error!("Failed to set local backing file length: {}", e);
-                    return Err(libc::EIO);
-                }
-                if !self.overlay()
-                    && let Some(sd) = self.staging.dir()
-                {
-                    sd.resize_bytes(old_staging_size, sd.file_size(ino));
-                }
-                if let Some(entry) = inodes.get_mut(ino) {
-                    entry.size = new_size;
-                    entry.mtime = SystemTime::now();
-                    entry.ctime = entry.mtime;
-                    entry.set_dirty();
-                    if new_size == 0 {
-                        entry.xet_hash = None;
-                    }
-                }
-                drop(inodes);
-
-                // Schedule flush so the truncation is committed to CAS/bucket
-                if let Some(fm) = &self.flush_manager {
-                    fm.enqueue(ino);
-                }
-            }
-        }
-
-        // Apply metadata-only changes (mode, uid, gid, atime, mtime)
-        if mode.is_some() || uid.is_some() || gid.is_some() || atime.is_some() || mtime.is_some() {
-            if self.overlay() {
-                let full_path = {
-                    let inodes = self.inode_table.read().expect("inodes poisoned");
-                    inodes.get(ino).ok_or(libc::ENOENT)?.full_path.clone()
-                };
-                // Reject any metadata-only mutation on a clean remote entry:
-                // there is no local backing to record it on, and the change
-                // would silently disappear at the next remote refresh.
-                let local_exists = self.local_backing_exists(ino, &full_path).map_err(|e| {
-                    error!("Failed to check local backing file for ino={}: {}", ino, e);
-                    e.raw_os_error().unwrap_or(libc::EIO)
-                })?;
-                if !local_exists {
-                    return Err(libc::EPERM);
-                }
-                if let Some(new_mode) = mode {
-                    self.set_local_backing_mode(&full_path, new_mode).map_err(|e| {
-                        error!("Failed to update local backing mode for ino={}: {}", ino, e);
-                        e.raw_os_error().unwrap_or(libc::EIO)
-                    })?;
-                }
-            }
-
-            let mut inodes = self.inode_table.write().expect("inodes poisoned");
-            if let Some(entry) = inodes.get_mut(ino) {
-                if let Some(m) = mode {
-                    entry.mode = m;
-                }
-                if let Some(u) = uid {
-                    entry.uid = u;
-                }
-                if let Some(g) = gid {
-                    entry.gid = g;
-                }
-                if let Some(a) = atime {
-                    entry.atime = a;
-                }
-                if let Some(m) = mtime {
-                    entry.mtime = m;
-                }
-                entry.ctime = SystemTime::now();
-            }
-        }
-
-        let inodes = self.inode_table.read().expect("inodes poisoned");
-        match inodes.get(ino) {
-            Some(entry) => Ok(self.make_vfs_attr(entry)),
-            None => Err(libc::ENOENT),
-        }
-    }
-
-    /// Read a file inode's fields into a `FileEntry` snapshot.
-    fn get_file_entry(&self, ino: u64) -> VirtualFsResult<FileEntry> {
-        let inodes = self.inode_table.read().expect("inodes poisoned");
-        let entry = match inodes.get(ino) {
-            Some(e) if e.kind == InodeKind::File => e,
-            _ => return Err(libc::ENOENT),
-        };
-        Ok(FileEntry {
-            xet_hash: entry.xet_hash.clone().unwrap_or_default(),
-            size: entry.size,
-            is_dirty: entry.is_dirty(),
-            full_path: entry.full_path.to_string(),
-        })
-    }
-}
 
 // ── VFS types ──────────────────────────────────────────────────────────
 pub type VirtualFsError = i32;
