+++ tests/csi_bidirectional.rs
@@ -72,19 +72,14 @@ fn get_plugin_capabilities_response_roundtrip() {
     let original = GetPluginCapabilitiesResponse {
         capabilities: vec![
             PluginCapability {
-                r#type: Some(plugin_capability::Type::Service(
-                    plugin_capability::Service {
-                        r#type: plugin_capability::service::Type::ControllerService as i32,
-                    },
-                )),
+                r#type: Some(plugin_capability::Type::Service(plugin_capability::Service {
+                    r#type: plugin_capability::service::Type::ControllerService as i32,
+                })),
             },
             PluginCapability {
-                r#type: Some(plugin_capability::Type::Service(
-                    plugin_capability::Service {
-                        r#type: plugin_capability::service::Type::VolumeAccessibilityConstraints
-                            as i32,
-                    },
-                )),
+                r#type: Some(plugin_capability::Type::Service(plugin_capability::Service {
+                    r#type: plugin_capability::service::Type::VolumeAccessibilityConstraints as i32,
+                })),
             },
         ],
     };
@@ -115,18 +110,14 @@ fn node_get_capabilities_response_roundtrip() {
     let original = NodeGetCapabilitiesResponse {
         capabilities: vec![
             NodeServiceCapability {
-                r#type: Some(node_service_capability::Type::Rpc(
-                    node_service_capability::Rpc {
-                        r#type: node_service_capability::rpc::Type::StageUnstageVolume as i32,
-                    },
-                )),
+                r#type: Some(node_service_capability::Type::Rpc(node_service_capability::Rpc {
+                    r#type: node_service_capability::rpc::Type::StageUnstageVolume as i32,
+                })),
             },
             NodeServiceCapability {
-                r#type: Some(node_service_capability::Type::Rpc(
-                    node_service_capability::Rpc {
-                        r#type: node_service_capability::rpc::Type::GetVolumeStats as i32,
-                    },
-                )),
+                r#type: Some(node_service_capability::Type::Rpc(node_service_capability::Rpc {
+                    r#type: node_service_capability::rpc::Type::GetVolumeStats as i32,
+                })),
             },
         ],
     };
@@ -145,10 +136,7 @@ fn create_volume_response_roundtrip() {
         }),
     };
     let decoded = roundtrip(&original);
-    assert_eq!(
-        decoded.volume.as_ref().unwrap().volume_id,
-        "hf-vol-test-123"
-    );
+    assert_eq!(decoded.volume.as_ref().unwrap().volume_id, "hf-vol-test-123");
     assert_eq!(decoded.volume.as_ref().unwrap().capacity_bytes, 1_073_741_824);
 }
 
@@ -158,17 +146,16 @@ fn node_publish_volume_request_roundtrip() {
         volume_id: "vol-42".to_string(),
         target_path: "/mnt/data".to_string(),
         staging_target_path: "/var/lib/kubelet/staging".to_string(),
-        volume_context: std::collections::HashMap::from([
-            ("source".to_string(), "repo openai-community/gpt2".to_string()),
-        ]),
+        volume_context: std::collections::HashMap::from([(
+            "source".to_string(),
+            "repo openai-community/gpt2".to_string(),
+        )]),
         publish_context: std::collections::HashMap::new(),
         volume_capability: Some(VolumeCapability {
-            access_type: Some(volume_capability::AccessType::Mount(
-                volume_capability::Mount {
-                    fs_type: "nfs".to_string(),
-                    mount_flags: vec!["vers=4".to_string()],
-                },
-            )),
+            access_type: Some(volume_capability::AccessType::Mount(volume_capability::Mount {
+                fs_type: "nfs".to_string(),
+                mount_flags: vec!["vers=4".to_string()],
+            })),
             access_mode: Some(AccessMode {
                 mode: access_mode::Mode::SingleNodeWriter as i32,
             }),
@@ -177,6 +164,6 @@ fn node_publish_volume_request_roundtrip() {
     };
     let decoded = roundtrip(&original);
     assert_eq!(decoded.volume_id, "vol-42");
-    assert_eq!(decoded.readonly, true);
+    assert!(decoded.readonly);
     assert!(decoded.volume_capability.is_some());
 }
